[roll] Roll fuchsia [fshost] Set minfs-max-bytes correctly.
When minfs-max-bytes is not needed for a configuration, it should be
omitted from the fshost configuration. Currently it's being set to the
empty string which is giving errors in the syslog at boot.

The issue was that the name of the argument from the invoker in the
fshost config template was actually catching the build parameter. The
result was the default configuration looked like it was overriding the
value with "". This change just uses a different name for the invoker's
override parameter.

There should be no behavior change since this set properly when needed,
and it was ignored when set to empty. This just removes a stray warning
in the syslog.

Original-Bug: 77049
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549901
Original-Revision: 6477e58450efba387f727f3b352f5bdb85308c37
GitOrigin-RevId: 80ca4b2585192201b91d19048e0f2ff97b375f11
Change-Id: Ifbbcae909625504b2ace82154fc6749b2b63ad35