[roll] Roll fuchsia [diagnostics][tests] Fix flake in cpu stats
It turns out that appmgr can have a measurement taken when it hasn't
been in the queue at all. This is likely because it is one of the only
things running at system startup, so it doesn't need to wait for CPU
time in the scheduler queue.

This CL fixes this test where it would fail due to a 0 queue time for
appmgr.

Original-Fixed: 80020
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551161
Original-Revision: 07b64a7f324dcb872070505f86fa4e83bc53e633
GitOrigin-RevId: 3628e6fa09a14a666c47e11353274f5acdb04243
Change-Id: Ib627431e407c3f11a7cda8f30205c3c950dd212f