[roll] Roll fuchsia [netstack] Fix test_case usage in DHCP
The `test_case` macro doesn't assert on a Result returned from the test
body unless the result is specified with `=> foo`. Since we
test_case-ified the DHCP tests with a Result return value, they have
been silently passing.

This change modifies the tests to not return a result and also fixes a
bug in the tests that was masked by this behavior.

Updates: 73252
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550604
Original-Revision: 61ddd63f40f05ce1732acdfd4a8f50672449e385
GitOrigin-RevId: 789c244993b839b5ca7eeb428bf935ec376160c5
Change-Id: Icf349fcf7aa7f6751935e0951808402db7b3697c