[roll] Roll fuchsia [pkg-resolver] Remove 1xx range status scan.
In testing the status ranges, it was discovered that tests can flake due
to hanging in the 1xx status range scan when connections are created
faster than they're removed from the pool. A comment was added to this
effect in all the tests. The comment indicates that flakiness is
experienced when requests come in faster than connections are removed
from the pool. This is referencing a pathological case where the hyper
HTTP keepalive pool provides a dead socket to the client, which does not
make progress on the request.

Indeed, `pkg_resolver_create_tuf_client_status_ranges` also inherited
this comment, but was nevertheless committed with this status range
scan. It is possible that this is contributing to the flakes mentioned
in the associated bug. Removing the scan for this space in any event
since, even if it's not the root cause for this flake, it is known to
contribute to flaky behavior.

Original-Bug: 77864
Multiply: pkg-resolver-integration-tests#meta/metrics.cmx
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551181
Original-Revision: c64c0846e96ad79ca9014eaafe46e8a59b1a1fd1
GitOrigin-RevId: bfc31bedc553402b5c2ecef0b0e648f727a64ba5
Change-Id: Id7fdc6d4a1e142a03de3917df54715267a4da61b