[roll] Roll fuchsia [bt][lib] Add Vigil helper library
Vigil is a library which lets you add actions that are queued to happen
when the item is dropped. It defines a trait, DropWatch, which takes a
function to be called when the item has been dropped.

To be used in A2DP and HFP profiles.

Original-Bug: 64521
Tests: added unit tests; fx test vigil-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548601
Original-Revision: 994dca45cfb8ed850a6e57b399440548ef7300ed
GitOrigin-RevId: e906553909ff01436b96129639fba65b2c54cef0
Change-Id: I47e8e2b30cd81399c2b3fc4fd664708c8a8bea4f