[roll] Roll fuchsia [test][docs] Update v2 test examples
Now that RealmBuilder is available, add a reference to the RealmBuilder
page.
This also replaces the example that used the driver pattern as Archivist
no longer needs to be manually placed in a test realm. Instead, this
adds a simple example where the test connects to a sibling component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547904
Original-Revision: 1fa9a397d81d98fa439cb43049c035b3b01efb13
GitOrigin-RevId: df849c7a16c8c1f204b1ede837bcca7665a36014
Change-Id: If8b57ebf3fe566d948ba7db2d5af9d622caf0785