[roll] Roll fuchsia [netstack] Support route preferences
This change adds support for routes to be sorted based on preference
and for the NDP dispatcher to add routes with some preference value.

Routes to the same destination will be sorted based on their preference.
If the destination and preference are the same, then the metric value is
used to break the tie (then following the existing sorting algorithm for
routes).

Original-Bug: 79015

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548354
Original-Revision: 7f42cda935d49de8a83d4a4ae929d2b138eb22da
GitOrigin-RevId: bf40bdaa44ecf2171a6b07fa1a7d06dd6d5fcf8e
Change-Id: I3a826dbf9553a20d56685c1acf72add10748ed77