[roll] Roll fuchsia [ffx] Handle non-existent files in the sdk manifest.
Make the File handle Optional, and if it doesn't exist return empty
data. Previous behavior is that if some file in the manifes.json is
missing, then the sdk plugin fails. We should want the parser to be more
resilient, instead of erroring, we'll set empty data to the manifest
field.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551380
Original-Revision: abd6168c86a551ed27e7caf237613033fce472e4
GitOrigin-RevId: 9b4d970c4fa9a9f7d02950ca0ebeeda78edf0a9d
Change-Id: Iba3c16fa2d09addbd44691f1db12a32c2041a0c7