[roll] Roll fuchsia [zircon] Add ZX_EXCEPTION_STATE_THREAD_EXIT
Currently, userspace can either handle an exception or have the
exception continue to propagate. This CL adds the ability for userspace
to make an exception cause the thread to exit.

This mechanism will be used by starnix to terminate threads in the child
process. When a thread in the child process calls the Linux exit() UAPI,
Starnix catches the BAD_SYSCALL exception and marks the exception as
FATAL, which causes Zircon to tear down the underlying Zircon thread in
the child process.

Test: Threads.ExitViaException
Original-Bug: 77376
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550383
Original-Revision: 5bb5372ce89d829b28da21d5e644560d67da934c
GitOrigin-RevId: 18586b86f43161d62a7ef33ab86c2711293be64c
Change-Id: I624306fcffc07c05b6ae7ad4b492ee76c9ede6fd