[roll] Roll fuchsia [fx repro] Read fint_params_path from properties
This is more robust that constructing the path using the builder
project/bucket/name, which may break in the future if, for example, we
ever have multiple builders use the same fint params file. The existing
strategy already doesn't work for subbuilds and rerun builders, which
have names that differ from the fint params files they use.

This relies on recipes change fxrev.dev/551763, but is backwards
compatible with older builds. We'll probably want to keep the backwards
compatibility around for a year or so.

Original-Bug: 80137
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551819
Original-Revision: e8dee622fb79027ffe151990a16ad6e3d4567795
GitOrigin-RevId: 21d205c63b953fa0ea99a5c7ff67546f3d821708
Change-Id: I2b0dab2b994747f7d2fc7c089c6e184328c6daf2