[roll] Roll fuchsia [zbi] Fix compression control flow
It turns out that ZSTD_compressBound() does not actually return an
upper bound on the size of the would-be compressed data: its
documentation reads
```
When there are a lot of small blocks due to frequent flush in streaming
mode the overhead of headers can make the compressed data to be larger
than the return value of ZSTD_compressBound().
```

The compression logic previously assumed that an output buffer of that
size would be sufficient - and a counterexample was met in the wild,
resulting in a crash. In this change, we update that logic to
iteratively request more output buffers as they are filled.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551887
Original-Revision: 57b6add15949fbef0993f20ad97546e6f4f12e79
GitOrigin-RevId: e9949d96194607d464fb8deeac9fe6efa149b2c4
Change-Id: I9ebf733d07f658c1f3db3435b40d65c9e63dbc08