[roll] Roll fuchsia [x86][acpi] Sketch AcpiEvaluateObject over FIDL support.
This CL does all the plumbing within the x86 board driver that's
necessary to be able to call AcpiEvaluateObject from a driver.
It also implements logic to check that a call is on an object that's the
child of a device.

Original-Bug: 79172
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551070
Original-Revision: 7756df3d1a46847442aa5650147e8ca98681039c
GitOrigin-RevId: c04e5a9cb617ab58b06f6bf6312aa632103a4a4e
Change-Id: If2a3a3c4461eb95ed1e61e60f590469d60d6bb7f