[roll] Roll fuchsia [ffx] exit with non-zero status if experiment not opt-ed in
This changes ffx to exit with a non-zero status if someone tries to run
an experimental CLI, but did not opt-in to the experiment. This is
useful in scripting, since otherwise the script would assume the command
succeeded.

Original-Fixed: 80069
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551366
Original-Revision: b16a58607cfae8a1f528bdd01169a2eaaf76fa9e
GitOrigin-RevId: 6e549876318b8d636ad9bf3194945b49818de2ce
Change-Id: I57a560f9e527f4b4931bcd198e1db3d59ea50798