[roll] Roll fuchsia [fidlc] Fix dangerous identifiers in new syntax
Tests for `true` and `false` have been turned off, to be revisited by
fxbug.dev77561.  The resolves an edge case in the new synytax:

```
service Foo {
  bar client_end:true; // Should this be a literal 'true' or constant?
}
```

In the future we'll need to decide whether or not the `true` and `false`
constants may be overwritten (in which case they should always be parsed
as identifier constants), or whether such shadowing should be banned by
the compiled.

Additionally, other common identifiers are now tested as in the case
of constraints.

Original-Bug: 77561
Test: fidldev test fidlc
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551816
Original-Revision: f0143c1f128f14610b4ebcc6de42e3833a67170c
GitOrigin-RevId: 4cb29fdc2e9fac3a614817df8f554bd009183f40
Change-Id: Ib1188e3ef719ca6bd66338c572aa5f51ca5f760f