[roll] Roll fuchsia [fidl][rust] Validate that envelope byte count is a multiple of 8
This CL makes FIDL decoding Rust validate that the byte count field of
all envelopes is a multiple of 8. Previously this would trip an
assertion in the decoder, making it possible for a client to crash a
server simply by crafting a FIDL message in this way.

This also adds a GIDL test, so it's now tested for all bindings. It is
currently denylisted in Go, which returns the wrong error code.

Test: fx test fidl_rust_conformance_tests
Test: manually checked that new tests failed before fixes
Original-Bug: 80162
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551921
Original-Revision: 1a99c907aab7e63cd902dd1f1d7c63a4dd0fa989
GitOrigin-RevId: 1304e52896adf101af5d657edbb4a076a6e0d5ac
Change-Id: Ia3ab560ab11c8e19e6e81a9bed7153ba251e7eb1