[roll] Roll fuchsia [lib/arch] Establish cache operation naming conventions
Whether the operation is for the local CPU or all of them is now
explicitly qualified with a qualifier of "Local" or "Global". (This
convention was previously present in part.)

Moreover, `cache_way_set_op()` is renamed to `data_cache_way_set_op()`
to clarify what cache it is operating on.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552533
Original-Revision: 41ef5f91441c6642b67fa878857fbc044ae55589
GitOrigin-RevId: 26494e26313f275c29b450532c7e50e2a5e4c98c
Change-Id: Ie87cdb9eae03b676378df13f7e48d1cd6154172a