[roll] Roll fuchsia [debugger] Add cross-platform status.
Currently, the status values have some undesirable properties:
 (1) Some status values are returned via IPC as zx_status_t constants.
     We avoid using zx_status_t in cross-platform code, so these are
     integers. Then the frontend has to interpret these values.
 (2) Sometimes, it's desirable to include a better error message with a
     failure.
 (3) If we port the debug_agent, zx_status_t will not be usable.

Adds a new cross-platform Status structure to deal with these problems.
A zx_status_ and an errno interface are provided. These are
self-describing and also allow custom error messages.

The intent is to replace all zx_status_t in cross-platform code with
this class and also serialize it to the frontend. This currently just
changes the ProcessHandle and its users as a proof-of-concept. When all
users are complete, we should be able to remove the calls to
Status::platform_error().

This uses the "debug" namespace. "debug_ipc" is ugly and annoying to
type, and doesn't even make any sense in the shared/ directory. We
could have "debug_shared" or "debug::shared", but that seems
unnecessarily long. I'm thinking we should convert everything in shared/
to "debug".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551889
Original-Revision: 50abb4de95f9effb70c2eccc9603cd9577da148b
GitOrigin-RevId: e9de9eb35367b4a45079f28614fd7c312d1ebb9e
Change-Id: I8967b77ad0deabf142a66c06ef4d7f1a2e1e6102