[roll] Roll experiences [ermine] Allow dismissing screen saver on keyboard activity
This change fixes a corner case of dismissing the screen saver when a
child view is running.

The screen saver can be dismissed using pointer or keyboard activity.
But this does not work when a child view has focus, the shell does not
receive keyboard events then.

In order to receive keyboard events, while the screen saver is active,
this change moves the focus to the shell view by showing the overlays.
This ensures that we get keyboard events while child view is running.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/552640
Original-Revision: d91d900c9494451d0dccd394a50338327a638a49
GitOrigin-RevId: 035856bb1cc5bfd00e9cd7ec3d43d57becb3865f
Change-Id: Ib83d4f2ca17439610bc3ef41b6a3bb8d7052f1c5