[roll] Roll infra/recipes [rbe] Simplify directories/paths.
We only really need a single directory for RBE, the outputs will not
conflict across the directories passed to the tooling. Also use a
temporary directory so that we do not conflict across multiple
executions of RBE.

This also fixes bugs where we weren't creating the directories at all
which RBE requires and the server address was passed without a unix://
prefix.

Original-Bug: 80255
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/552821
Original-Revision: b53db3a7030f98699285ae0467ff37e3cd3b1462
GitOrigin-RevId: b4181178a0f51ead809e7e01af0d731e5db6e053
Change-Id: Ie23bbf54a1b551d70b08617bdfdfd0b6f9487033