[roll] Roll fuchsia [blobfs] Make metrics used by pager thread-safe.
Adds locks to the fields that are flushed to cobalt to make their access
thread-safe. Marking them atomic was a consideration since they are
largely just incremented, but the zx::ticks are not so easily converted
to atomic. The inspect fields do not require locks for modifying as
those fields are already thread-safe as they are guarded by the inspect
tree lock.

Additionally adding locks for the page-in frequencies, as they do some
operations on the map that would not be thread-safe.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552441
Original-Revision: aefe7ca9912fa15643483c18ef87cce29f09b7db
GitOrigin-RevId: 1d4305b435e3f37c11bd2cd455461a77769aaece
Change-Id: Id888405b9414f652b3c0e97f01859c9cd211558b