[roll] Roll fuchsia [fint] Add fvdl_sdk_test to never-affected tests
This is an e2e test that depends on the fvm so it's technically
affected (at least according to the build graph) by a large fraction of
fuchsia.git changes despite the test not actually being sensitive to
most such changes. So it's not useful to automatically multiply it on
those changes, and considering affected also makes it more likely that a
build will exceed the maximum of 3 affected tests allowed before no
affected tests will be multiplied in a build.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552982
Original-Revision: 20245eeec797c852c9d1d31c3ba5904b2845490c
GitOrigin-RevId: 22203a26de7507fd4405927ac35cb05053b0b124
Change-Id: I8f2fd164329ff181c64f85111fed5dd538444c0b