[roll] Roll fuchsia [mock_ddk] Convert ethertap driver tests to mock_ddk
This is part of a set of changes to discontinue use of the fake_ddk.
New drivers are instead recommended to use the mock_ddk or depend on
//src/devices/testing/no_ddk to get the necessary functions linked in.

This change also should address a flake with this test.

Original-Bug: 80142

Testing: This is a change to test code.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552682
Original-Revision: 2de6d21992bc91d9dc47b78718d888609a2c40ef
GitOrigin-RevId: a30528e1ec5f2ea6a91ab61fd32bbb22b0348b79
Change-Id: Idef71bf8412a22ea666eb0acbc63276f5e1e54cc