[roll] Roll fuchsia [ffx][repository] Register targets on start
This changes `Repo::register_target` to save which repositories we've
registered to config. Then in `Repo::start`, read the config and
re-register the repositories on the target. This allows ffx to make sure
that if the daemon is restarted that it makes sure the repository is still
set up.

In addition, this:

* Adds RepositoryManager::clear in order to drop all repositories when
  stopping the server.
* Changes FakeDaemon to support specifying the target nodename.
* Added a test to connect to the default target.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551308
Original-Revision: eefe402cba963ef0d57ab6d62c70bdfa914db653
GitOrigin-RevId: d1ea196d0112953f79c3295474074ce82d02e0a4
Change-Id: I754cfe10b7a4799a256b438159c3a6eec98ae7b9