[roll] Roll cobalt [EventLogger] Fix duplicated logging
Cobalt 1.1 would log to each report a number of times equal to the
number of reports in a metric. This change fixes that, so that each
report is logged to only once.

Additionally, found bug in implementation of UNIQUE_DEVICE_HISTOGRAMS,
and UNIQUE_DEVICE_NUMERIC_STATS, where the values would be aggregated as
hourly, instead of daily.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/552530
Original-Revision: 4a681cc5427f45144e12c8ff8e91900f3dccb511
GitOrigin-RevId: 229deb772464f010e2127155736b0c308605f301
Change-Id: I3d7873f6e5227ded95ff67b9ab4cd93404546a43