[roll] Roll fuchsia [fostr] fix fb/79996 and fb/80062
This change fixes a couple of bugs, one of which was a regression.

Nullable fields whose value is defined in another library were not
being formatted properly. This is due to the way operator overloads
for std::unique_ptr were handled. To avoid defining an overload for
all unique pointers for the world, a overload template appeared at
the top of every formatter.cc file that formats unique pointers for
the library namespace. This works ok when the nullable value is in
the current library, but doesn't kick in when the value comes from
another library (we get "0" for null or a memory address in hex).

The other issue was a regression introduced when HLCPP started using
std::vector. To avoid defining an overload for std::vector generally,
vectors were handled using a PrintVector function that returns a
string. This didn't work correctly, because the ostringstream used
to produce the resulting string didn't have the indentation context.

Both issues have been fixed by introducing a disambiguating wrapper
template fostr::Formatted<T>. This wrapper can wrap anything (by
reference) and allows us to write overloads for e.g.
fostr::Formatted<std::unique_ptr<T>>, which cannot be confused with
vanilla std::unique_ptr<T>. The wrapper is required for all std
namespace types (std::unique_ptr, std::vector and std::array), so
the codegen template has been changed accordingly. The
unspecialized version of the wrapper just formats the wrapped item,
so fostr::Formatted can be freely applied to any value without an
issue.

Original-Fixed: 79996, 80062

Test: added unit tests for regressions

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551883
Original-Revision: 727de440e28531f7176a8256cbd698132015ebf6
GitOrigin-RevId: c2ad579a8cf54de19d43eceeadd72ffbc5963d56
Change-Id: Ia830e6c1278f9f984a4de6416cfe599fee71f9f6