[roll] Roll fuchsia [docs] Improve event capabilities docs
A lot of this will be rendered obsolete by the RFC
(https://fuchsia-review.googlesource.com/c/fuchsia/+/535692) currently
in review. But this represents the current state of the world and the
state we'll be for at least a few more months.

Original-Fixed: 73003

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547915
Original-Revision: 2efa325a15cfd5e696ba67373a9508b16825e101
GitOrigin-RevId: 2202a1ad647ccfb511ca9d224ce2f0af6ee67380
Change-Id: I4e42d53e0f70a55e137291d76476e0a1a7869f22