[roll] Roll fuchsia [vmm][arm64] Always report that the UART TX FIFO is empty
PL011 gives separate TX FIFO "full" and "empty" bits; currently the
hypervisor always reports that both are false - so, in particular, that
the FIFO is never empty. It is a reasonable expectation for any driver
that this FIFO should eventually be empty - and we have a platform
driver that does expect this. Accordingly - and since incoming data in
the emulated UART is immediately processed - this change updates the
hypervisor to always report that the TX FIFO is empty.

Original-Bug: 80376
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553322
Original-Revision: 69698d9be99e141037c9947115211b64bad419c4
GitOrigin-RevId: 54b75960b484c430618876cc4a3c910e99546886
Change-Id: I6fc065abfa1169cf9f27a357a6f9e8a4d1ebd4b5