[roll] Roll fuchsia [starnix] Use OpenFlags more
Now that we have a type for OpenFlags, this CL adds more declarations to
the bitfield and uses the type in FileObject.

This CL also refactors open_internal into open_file_internal and
open_node_internal. The former takes OpenFlags and returns a FileHandle.
The later does not take OpenFlags and returns the underlying FsNode. The
CL updates the callers to use the appropriate entrypoint.

Test: No behavior change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553382
Original-Revision: 5cfd27d12e8748d1702e493c0650b36962e33a53
GitOrigin-RevId: c94f182265bcec0331409fdd284df54f0357ab17
Change-Id: I2287e6458e8796a9df6b14172f0ca434cd372736