[roll] Roll fuchsia [starnix] Restore signal mask properly
Previously the signal mask was restored when the task woke up inside of
sigsuspend. This would cause the signal dispatching logic to check the
signal against a different mask (the restored one) than was used to
determine whether or not to wake the task in the first place.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553146
Original-Revision: c827a116636b8fb0369e29b1779cb4a58e12e2c3
GitOrigin-RevId: 25a828ea1e0f3b45581270fd64c68562225af26d
Change-Id: Ie5d7310334a6558e5d180c2380c2671e9e5fb0c9