[roll] Roll fuchsia [fdf][composite] Add primary_fragment field.
To align with the future change, where one of the fragments will be
identified as the "primary" fragment. A future change will replace the
`coresident_device_index` with a boolean which describes whether or not
to load the driver which binds to the composite device in the same
driver host as the one hosting the primary fragment.

Some of primary fragment choices are a bit awkwards, and a // ???
comment is present to denote them.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551896
Original-Revision: 49e667be9bde084474c2f7ac4b0898330f9ea039
GitOrigin-RevId: 0d79794e3c65309e493d4792caf7f7314715ce90
Change-Id: I73a0630c205279722372ccbeb4448c767e3a94f3