[roll] Roll fuchsia [debugger] Add function step controls.
Generalize what to do when a new function is encountered during
stepping. This gives us a place to add logic to skip library functions
in the future.

Separates out the behavior for unsymbolized code from PLT trampolines.
It now always steps through the PLT code, while the unsymbolized
behavior applies to regular functions. The unsymbolized handling has
moved from a flag on "step" to a setting.

This replaces an old hack that stepped through the PLT trampolines by
detecting whether we were in unsymbolized code in a module that seems to
have symbols. We have since added real ELF symbols which allows us to
check for this case directly since the ELF PLT trampoline will have a
correspondingly annotated symbol.

Update the StepThreadController test to set up the symbols properly for
this test. Use the new InjectExceptionWithStack() to control symbols
more precisely.

This new behavior is incrementally better and is more flexible for
future additions. But it will still need to be significantly enhanced to
properly handle the general PLT problem (discussed in issue 80275).

Original-Bug: 80258
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552281
Original-Revision: ba1f386b590e46f1aa208946ca2fd96df1fd325a
GitOrigin-RevId: c99d85d1913e4a7692c70aa0458a7fc49d0a33bb
Change-Id: Ib4a4666d4b2c7666652c6b08444e5993165d61e8