[roll] Roll fuchsia [fdf][pbus] Replace coresident_device_index with primary_fragment
This was already updated in the libdriver API, this change just
propogates the change to the pbus API. One notable difference is that we
specify nullptr to specify that the primary_fragment is pdev.
Additionally, unless the primary fragment is set, we always spawn the
driver in a new driver host.

There should be not functional changes from this commit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551899
Original-Revision: ae364f3a403bd7ad689c174d72aec6d2ba943c3d
GitOrigin-RevId: 5ff8a40d7b7af7b19961e3b75253268f77c58f85
Change-Id: Ic2b20a3594567110c749dfc0199fccfc10e2d47d