[roll] Roll fuchsia [netemul] Return unescaped moniker
...from fuchsia.netemul/ManagedRealm.GetMoniker. Currently GetMoniker
returns a moniker with the ':' character escaped with a '\', which is
needed to use the moniker as a selector when reading inspect data.
Instead, return the actual moniker, and when using the moniker as a
selector, use a diagnostics utility to escape the necessary characters.

Test: fx test netemul-v2-tests
Test: fx test netstack-integration-tests-v2
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553307
Original-Revision: 83af9636a642481f11b0946405b7817b7696aafa
GitOrigin-RevId: 402fbbe45631e6edd4e55b589de4dde90b70ac81
Change-Id: I7def114e2d4a8a82e879e04b18796fec10d5991f