[roll] Roll fuchsia [mock_ddk] Convert aml-nna driver tests to mock_ddk
This is part of a set of changes to discontinue use of the fake_ddk.
New drivers are instead recommended to use the mock_ddk or depend on
//src/devices/testing/no_ddk to get the necessary functions linked in.

Testing: This is a change to test code.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552029
Original-Revision: 13f76a1bd0c7bbc451c0d2a38ed3a4c80c31f68a
GitOrigin-RevId: a5f1fe770defaf1aa0172c07b665a62151ceea84
Change-Id: I22300fb27e2758ceeaf80df636afbb4ccda6f72e