[roll] Roll fuchsia [fdf] Add option to reboot on driver crashes
For products where we aren't confident in the system's ability to
recover from driver crashes, we would like to reboot the system instead
when we observe a crash.

The currently implementation just crashes the driver manager in response
to a driver host crashing, which will trigger the system to reboot due
to the fact driver manager is marked as a critical component. This can
be updated to do something more graceful in the future.

devmgr-integration-test was updated to support exposing the process
handle for driver manager, as well as share the exception info with the
exception callback registered.

Original-Bug: 67168
Tested: fx test isolated-devmgr-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552990
Original-Revision: 2e2025a71b922d236d52e9e39d4957e3f800e433
GitOrigin-RevId: 5f76db954545a61f660ad2298a2ecf965778b1a5
Change-Id: Ibae3ee354aa76012f88db3bb509f375e2f01ad3b