[roll] Roll fuchsia [blobfs] Apply deadline profile to pager threads.
The switch to the new pager for blobfs lost the deadline profile on the
pager threads. This applies the deadline profile to all pager threads.

Test: manually verified that the function is called and succeeds.
Unfortunately, there is currently no way to query the profile on a
thread such that we can ensure this going forward.

Original-Bug: 80357
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553209
Original-Revision: 74f311184bc4695a545891950b982e1a700dbd8f
GitOrigin-RevId: 48221e4477a63e30f079e112db99686dad490dc1
Change-Id: Ic027e62fea841195ce78f43a078bbe82d8fa61ad