[roll] Roll fuchsia [debugger] Print return values before a stop.
Previously, the function return value from "finish" was printed before
the stop notification if it was synchronous (say, in a register), and
after if it was asynchronous (it was on the stack or printing it
required fetching memory). This is because the thread would continue its
work (either stopping or continuing) in parallel to the value being
evaluated.

This also prevented showing more return values that might happen in the
middle of stepping, say in "next".

This patch adds a system to evaluate these asynchronous expressions
before resuming the thread. It adds a new asynchronous phase before
resumption that avoids making all of the thread controllers
asynchronous. This allows the return value evaluation to have time to
finish before notifying the user or resuming the thread.

Original-Bug: 5720
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553101
Original-Revision: 94d10c1bf38bc9d4af8c1d05c68acd1130768d1b
GitOrigin-RevId: ff16f8b9b1e079a515c73b662e55b4cbe6174335
Change-Id: Ic4108eb663765457b8171c7e7d5d1ab5e6be855f