[roll] Roll fuchsia [scrutiny] Add ffx scrutiny list packages.
This change adds a new command to the Scrutiny FFX interface which
allows you to list all of the fuchsia packages in the current build.
This is intented to be used as a discovery mechanism for the the
ffx scrutiny extract package command which allows you to extract a
package from a given package url.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551843
Original-Revision: de0dfece13242072658cbe1b3054e239ca8f02fb
GitOrigin-RevId: fc10d60d3045b18c8757bae6f4aa190556f30d95
Change-Id: I64c28a2995d7dcd0179ff01cf195360dd3fd28e8