[roll] Roll fuchsia [helpdoc] add flag to escape jinja tags
Jinja tags in fx help docs cause a parser error in fuchsia.dev
publishing pipeline because the tags are attempted to be interpreted.
This CL adds a --escape-jinja flag that uses a mix of jinja escaping and
html escaping to avoid a parser error in includecode widget, which
expectes an argument and is validated for syntax issues before the verbatim
tags. Simple HTML escaping doesn't work because the reference docs in
fuchsia.dev is in a fenced code block, so the html entities would show
up as its code, not the expected symbol.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554197
Original-Revision: 32fc64227cf3ab801e2521f1dd62f8bd3cfb81ef
GitOrigin-RevId: c1ed0e67a7a9897cf15b78f126bad079d9129bec
Change-Id: I382a12180fcd87fb64e9ab86e8423db546897655