[roll] Roll fuchsia [starnix] Pass some of kill_test
This change passes all the signal tests that were blocked on bugs in
starnix's signal handling, as opposed to missing features (e.g., PID=1
workaround for missing /dev).

This fixes a bug where if a task was `sigsuspended` with a pending
signal, it would go to sleep and never wake. This would happen if the
`kill` came in during the `sigsuspend` execution.

It also adds support for waiting for a child process to exit (although
it does not encode the exit code correctly, yet).

Original-Bug: 79289, 76976

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554192
Original-Revision: 2016a0b048f387e78eda81dfcc2959d73c86be8d
GitOrigin-RevId: adeb9ebabc1062bcf833f45afd853c2bec8793b7
Change-Id: Iad149db176f80e261887390e66e62e45229662a7