[roll] Roll fuchsia [scrutiny] Add command ffx scrutiny list components
This change lists all of the components that scrutiny found in the build
for this particular image. This is useful for auditing when you want to
analyze a static build image to determine what components are in all the
packages.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552621
Original-Revision: fbfc687c2a432e2daaf950133d6f04338d26096d
GitOrigin-RevId: 33131e4f03f6bab914f33ef5d8d53c6c36d3b352
Change-Id: Ia80e5d4d05e8c18084178b7368e67e05ce865e24