[roll] Roll fuchsia [bt][mock-piconet-server] Add optional bt-rfcomm to topology
Provide API endpoint to optionally specify bt-rfcomm as an intermediary
component in the mock piconet. This matches production code in which the
RFCOMM component sits between bt-host and profiles.

Make a note that this should be used for integration tests of profiles
that require RFCOMM, and that integration tests for the bt-rfcomm
component itself should not use this hook.

Test: fx test mock-piconet-client-v2-tests
Original-Bug: 79742
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553045
Original-Revision: e5f7ea145729c5e70954e4a59a49b390cf33cadb
GitOrigin-RevId: 1b5400a5f88ad318ed3d79565cdbf3e4e967d5d6
Change-Id: I9ef91207cc80236d47ec160ba12c7cf39fbbb1cb