[roll] Roll fuchsia [kernel][boot-options]: kernel.serial.
There is some 'shared' logic between physboot and kernel proper now.
The logic is 'whether i found an entry for kernel.serial' in the command
line argument.
This is captured as 'gBootOption->serial_source' which is available
until full unification is achieved.

The kernel then will use this information to either parse something from
the commandline or the zbi, with its own logic.

This is necessary because Boot Options, at this time, does not provide
any information on whether the value is default, zbi, or command line.
So the kernel could not make the decision whether to fall back to the
zbi or not. Now it doesnt fallback it just uses whatever physboot
used(nothing, commandline or zbi).

Original-Bug: 53594
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546582
Original-Revision: c7e7bfb737c5c5d33db2e65a35ab1403054e57cb
GitOrigin-RevId: 242ef886011cf09ce3540c7541a49ef3195ebe1f
Change-Id: I382448892765755be9f2c72aa1bf04a10a11e7d3