[roll] Roll fuchsia [virtualization] Improve output of PeriodicLogger.
Make sure the "Finished" message of the previously logged item is
written out before the new logging item. Currently, this is not always
the case because we stop the logging thread _after_ printing out the new
log message.

Additionally, round the display "Finished after XXXs" to the nearest
second, to avoid values such as "59.99s" being confusingly displayed as
"59s", such as when the deadline was taken just before the log message
was started.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553941
Original-Revision: a6f67888708f0122bd6170d87260cee24946e2ba
GitOrigin-RevId: b511899e68d194d5fe69a9ea94fcbd955365781f
Change-Id: I246db9399147bb0385be0dfc0a5c891e9026c1b2