[roll] Roll fuchsia [flatland] make OnError separate
As part of the transition from OnPresentProcessed() to
OnNextFrameBegin() semantics, we should make the Error portion of the
event separate, as errors are on a per-present basis, and
OnNextFrameBegin will not necessarily fire 1:1 with previous presents.
It also helps simplify the logic conceptually for clients.

Test: changed unittests
Original-Bug: 79131
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553310
Original-Revision: 47ee7f4fb7130c7d8aeb4de1f239c71186ee273d
GitOrigin-RevId: 2b2a815aa1557501163d86902888948a0611f658
Change-Id: I000ea34d40e1c160d43fffea6827ddf171a0554c