[roll] Roll fuchsia [bt][avdtp] Have AVDTP commands send immediately
Update the API to send the command before returning the result, a Future
which you can poll to receive the response.

This makes it possible to send a command that you don't care to get the
result from (i.e. an Abort or a Suspend that should only result in an
acknowledgement) without needing to poll the Future.

Previously commands would also borrow the peer object and arguments
until the response was received. Because they are only used for
constructing the command to be sent, this also removes that borrow;
response Futures have unrelated lifetimes.

Test: added command_response_can_be_discarded; fx test bt-avdtp-tests
Original-Bug: 77017

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553970
Original-Revision: 4639d0e787eae98b3b85dc2d07e50f38a0151dfa
GitOrigin-RevId: 1fc210a221210d74a0b58b000462bc3f82905cce
Change-Id: I8358a667bea67ec3f94810e40643bdd2a29ba545