[roll] Roll fuchsia [network/tests] Only read logs from the netstack under test
Currently test_log_packets uses fuchsia.logger/LogListenerSafe to listen
to syslogs and verify the netstack appropriately logs when it sends/
receives packets according to its configuration. However, the
fuchsia.logger protocols don't include moniker information, and since
this test is connecting to the system-wide logs, it's reading logs from
every netstack that happens to be under test while the test is running.

This CL changes the test to use fuchsia.diagnostics/ArchiveAccessor,
which includes moniker information in logs. Then we filter the log
messages to exclude those not originating from the netstack spawned for
this test, so we correctly only pass/fail based on those logs.

Additionally, tests use the debug netstack, which is already configured
to log packets, so this test was not actually testing the
`fuchsia.net.stack/Log.SetLogPackets` knob. To fix this, this CL teaches
netemul how to override program args for components, and removes the
`--log-packets` arg from the netstack for this test.

Original-Fixed: 80036
Test: fx test netstack-fidl-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553966
Original-Revision: 15333c2d318a7a0af73e7cb95b3f18bfa77ea014
GitOrigin-RevId: 88aa183015b75d62814b77f831b332541acc46bd
Change-Id: Ic7146cc1450825184fe8702b0357c0432c97a34a