[roll] Roll fuchsia [docs][bt] Update persistent data schema doc
A variety of small cleanups so that the document accurately describes
the current schema used. Aside from small naming changes, elaborations,
or fixes to the examples, the most significant difference is the use of
a "PeerKey" type to represent a key with security properties, instead
of security-property-having and security-property-less Keys sharing the
same type, which affected a number of types which had Key components.

Original-Fixed: 79702
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553666
Original-Revision: d1ffc17cc183837b2eecdea61a032e87bfccbb12
GitOrigin-RevId: 095769b661638afb5c921ff7009dcdea9e8f51fe
Change-Id: I8ffa614b531b9f946b676f1f50c669d7af763428