[roll] Roll fuchsia [gigaboot] update logo
Update to the new Fuchsia logo, and clean up some of the logo code while
we're here.

Also adds a simple Python script to convert a logo image to RLE in case
we need to do this again in the future.

Original-Bug: 80398
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554785
Original-Revision: b705455ad773f7429c3f7c6760dd74e17cf459e5
GitOrigin-RevId: 8f58002b80790a88f209d35805617a0951a105e4
Change-Id: Ic0a9b28e23c5d6da5f51b8065a218480840712fa