[roll] Roll infra/recipes [rust] Add default-linker-libraries when building aarch64
Fuchsia compiles the rust compiler and tools from both x64 and arm64
hosts, but a recent upstream change added a dependency on the C type
`long double` that was never present before (apparently).

The dependency is from the `cargo` tool, using the `curl` C library,
which added the new type.

It is not straightforward to test this change without an Arm64
development machine, but the change does not appear to break when
building from an x86 host, so the risk appears low.

If this works, I hope to see the Rust canary builders for arm64 start to
work again.

The relevant bug ID is fxbug.dev/80522

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/554705
Original-Revision: ee95904b9272eb1f443b92d137bc7ce1bcedf49c
GitOrigin-RevId: ade8ee9c979455da70ac17d8ed92905fc042ebbc
Change-Id: I66eff64e794940e23400b2f1fb866894dc6b83f9