[roll] Roll fuchsia [build] Make clang_prefix not rebased by default.
This CL changes the definition of the 'clang_prefix' variable
to a source-absolute or system-absolute value (in the case where
it is redefined through args.gn). The fact that it was previously
rebased was a source of confusion.

It also adds rebased_clang_prefix which is the same path rebased
to root_build_dir, as it is used in multiple places.

For testing, the content of all .ninja files before/after this
CL were compared. The only difference is the path of the clang-format
tool used by the fidling_toolchain.ninja file which is not relative
to the root build dir, instead of being absolute.

Note: //third_party/go/BUILD.gn still uses rebase_path(clang_prefix)
which will expand to the same absolute path before/after this CL.
Fixing this to use relative paths instead requires a different CL in
the //third_party/go.git repository.

Original-Bug: None

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554562
Original-Revision: 17e77e6856a4bb24a5a2f2d7887d2c959b0df647
GitOrigin-RevId: 5f56f086f75515bd51a1fa47cdc07c1c41b8b8b6
Change-Id: I885d89521443c345a3aba9b46bba00232c0edd5b