[roll] Roll fuchsia [audio_core] Rename AudioClockManager -> AudioClockFactory
This is a nit:

Making this change because in VNext, I'd like to standardize terminology
and I've started using "Factory" for "thing that makes other things".
"Factory" is the standard term for this. A "Manager" typically manages
the lifecycle of objects which may be created elsewhere.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543101
Original-Revision: 2a602ae9a2cc5490f09b4155845af3fed2bb4768
GitOrigin-RevId: 3886bfc18cc46b7952938e29be4d58cd17cb492c
Change-Id: I31c536fb746bc964cfa829de7a8a2be8adb7962a