[roll] Roll fuchsia [virtualization] Make the test Logger class thread-safe.
Add a mutex to the Logger class, preparing to allow writes to the
singleton logger instance in future CLs.

The `Buffer` method becomes a copy, but is only called once on a test
failure, so the performance impact should be insignificant.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553944
Original-Revision: f9cf21fa477befe20e6744503855e9dc2ce1e613
GitOrigin-RevId: 6be7340357640ddadd3560b67e71cb83016ba23a
Change-Id: Ic2f8322801dbf311b6323bf9cd6007f295d554c1