[roll] Roll fuchsia [ffx] ffx e2e experiment tests
This creates an experimental plugin that can be called from an e2e test
to ensure that the code generation around experiments is working as
intended. We test that you get a non-successful exit code and a message
when you try to run a plugin that is a non-enabled experiment. And we
test that enabling that plugin allows you to run it successfully.

Fixes https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=80133

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554826
Original-Revision: 0b2684f3f80f966da17f95ef541f40988f7dc3db
GitOrigin-RevId: 73e9befebb1ab03792190e7c2df8c3d0e57f30e4
Change-Id: Icbe391ae94d36f60c2643349fd41bcf3efc63328