[roll] Roll fuchsia [docs] Update build-info v2 migration instructions
We will not support routing /config/build-info to v2 components, as this
perpetuates a deprecated solution for getting the build info at runtime.
We've been asked to move away from that feature, and have created an
allowlist of v1 components that use it. If we allow such v1 components
to migrate to v2 and keep that deprecated /config/build-info directory
then we will have to set up an equivalent v2 allowlist. This feels like
more trouble than it's worth.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553762
Original-Revision: e15ca6e4fa09c8f83f9077134a829239facb4bf3
GitOrigin-RevId: da01ea6505976cc9f9c2f4541f494bb980afb107
Change-Id: Ibf904a733f120371c4368bfef0d99f4785d95f9e