[roll] Roll fuchsia [kernel][boot-options]: Reland "kernel.serial."
This is a reland of c7e7bfb737c5c5d33db2e65a35ab1403054e57cb

There was the wrong assumption that ParseBootOptions was called only once,
but it is called once per zbi item cmdline. This means that the serial_source
needs to be preserved across calls.

This is done by using OptionSource::kDefault as the only state that can transition
to OptionSource::kZbi.

The original bug is as follows:
zbi cmdline item 1 : "serial=.... k1=v2"
zbi cmdline item 2 : "....."

When item 1 is processed, the serial source becomes kCmdLine, but when the second
one is parsed, it is resetted to kZbi, and because there is no serial option in the
second item, then the source is incorrectly set to kZbi.

Original change's description:
> [kernel][boot-options]: kernel.serial.
>
> There is some 'shared' logic between physboot and kernel proper now.
> The logic is 'whether i found an entry for kernel.serial' in the command
> line argument.
> This is captured as 'gBootOption->serial_source' which is available
> until full unification is achieved.
>
> The kernel then will use this information to either parse something from
> the commandline or the zbi, with its own logic.
>
> This is necessary because Boot Options, at this time, does not provide
> any information on whether the value is default, zbi, or command line.
> So the kernel could not make the decision whether to fall back to the
> zbi or not. Now it doesnt fallback it just uses whatever physboot
> used(nothing, commandline or zbi).
>
> Original-Bug: 53594
> Test: CQ
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546582

Original-Bug: 53594
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554821
Original-Revision: 85325767a22de6efa78a0a98b4ff7188fb22e9ba
GitOrigin-RevId: 940759830bd4e539183d184e52eb4dfeb92b6b78
Change-Id: Ic54e43b3a7acf047301f7966e49bcd06a461cefd