[roll] Roll fuchsia [fidlc] Fix C header generation error
https://fuchsia-review.googlesource.com/c/fuchsia/+/526761 missed a case
of kBox unwrapping.  Because the box is not unwrapped, the C generator
has no way of detecting that a method request/response argument is
actually a nullable struct.  The previous behavior was for such methods
to be silently dropped from the protocol.  An example of this can be
seen at sdk/fidl/fuchsia.sysinfo/sysinfo.fidl line 39 where the
`GetInterruptControllerInfo` method should be dropped.

The missed case caused the generator to ignore the "structness" of the
argument in the case of nullable structs, causing it to generate the
"droppable" method, which would in turn cause build time errors.  This
change fixes the oversight.

This behavior should be tested by golden tests, but no such tests exist
for the C generator output.  However, the post-conversion generated C
code has no diffs and builds successfully after this change, which
serves as a "golden" for the purposes of the conversion.

Original-Bug: 80353
Test: fidldev test fidlc
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554962
Original-Revision: 2d099f60fa4a860893a38d23014446e785fda14a
GitOrigin-RevId: a56f4f1943df1620e9f1e4fde2155ff863c78a4e
Change-Id: If68fff35cff84fc58fcc7476e4aa3a750ef0e4d0