[roll] Roll fuchsia [assembly] Add empty cache_packages to assemble_system.gni
The GN assembler includes a data/cache_packages file to the base package
even if it is empty. As we are unifying the assemblers, we need to
create this file in assemble_system.gni as well.

Original-Bug: 79886
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549570
Original-Revision: e81bcc053b3f699dc8c5af358e4f2b4e88d8e56f
GitOrigin-RevId: 565c47f8d8077d6db0b85d95a872696d6df8cf88
Change-Id: I57cb912150cc3d9cbaf9587ded1e58ee8d7ef234