[roll] Roll fuchsia [kernel][vm] Fix off by one error in enumeration
VmAddressRegion::EnumerateChildrenInternalLocked has an off by one in
its enumeration where it could eroneously call the on_mapping callback
for a VmMapping that was not in the range.

This is not normally triggered since the start of the iteration uses
subregions_.IncludedOrHigher to begin iteration at a valid mapping. But
if walking down into child vmars it is now possible to have mappings
that occur prior to the range be considered. The included test
constructs such a hierarchy as demonstration.

Original-Bug: 80034
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555763
Original-Revision: b95ae8a4e31731f4d157031fe45035ca1b8ba143
GitOrigin-RevId: c534905a3accd37a364d2cafe8659a4730c51a2d
Change-Id: Idac3b24e2eb962268832ba5c1272b915319f52eb