[roll] Roll fuchsia [assembly] Generate devmgr config in the image assembler
This is required in order to transition fully from the GN assembly code
to the image assembler. The devmgr_config.txt file is now generated
inside of the image assembler and added to the list of BootFS files. For
bringup (bootfs-only) builds, the devmgr_config.txt file is empty.

Original-Bug: 79674
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549563
Original-Revision: 8b2129b9fa666958a19cef10256b2e6aa2173f69
GitOrigin-RevId: 3d59b238d24b508e311b53026c220683b0fccf6c
Change-Id: I8745a179a84953fde3d9da18c5c9a86375731648