[roll] Roll fuchsia [componentsv2] (AbsoluteMoniker|Relative)Moniker support partial monikers
We would like to support binding with partial monikers that don't
include instance IDs as this is a runtime detail that isn't easily known
to developers. In order to do so we need support for
PartialAbsoluteMoniker and PartialRelativeMoniker. This CL introduces
both. We do this by making AbsoluteMonikerBase and RelativeMonikerBase
traits with associated types and implement concrete versions in
AbsoluteMoniker, ParitalAbsoluteMoniker, RelativeMoniker and
PartialRelativeMoniker.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553061
Original-Revision: 4533bf4e9ea0a4517539eee07a1480ebc6fb043d
GitOrigin-RevId: e999012f0958cc6bdbf148944aac12b0e941d57a
Change-Id: I822cd2cdaf3d72543906cc7a1ee974daee2a7e09