[roll] Roll fuchsia [packet] Improve ergonomics of FragmentedBytes
FragmentedBytes{Mut} used to have a single lifetime parameter for both
levels of references. However &mut T is invariant over T, this makes it
inconvenient to work with when we want the outer borrow to have a
shorter lifetime, transmute can be used as a workaround but after all,
it is unsafe. One more lifetime parameter is given to the structs, so
that they can be unified independently.

While I am here, anonymous lifetimes are removed in the where clauses
to not make readers confuse it with lifetime elision. Lifetime elision
in where clauses is not supported and its meaning is not even decided
(https://github.com/rust-lang/rust/issues/45667). To avoid surprises and
to be more specific, we ought to change those to use the explicit
desugared HRTB form.

Test: No behavior change
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552743
Original-Revision: 8b81a2baecd872709b86fd412433402f98b63e5b
GitOrigin-RevId: 0b7131281c578cb6b852d5d554901e4677e1b7ec
Change-Id: I1441844af55844ca6a75728706e1c8721eb79a56