[roll] Roll fuchsia [bt][sm] Abort upon unexpected opcode in Phase 3
While pairing, we only expect to receive certain opcodes, enumerated in
Phase3::OnRxBFRame. In other pairing phases, upon reception of an
unexpected opcode, we aborted the pairing process, but in Phase 3, we
only sent a Pairing Failed message. This could cause pairing to hang.
To fix it, abort pairing instead of just sending the message.

We also removed an invalid expectation in the security manager unit
tests - the expectation says that "receiving a security request now
should have no effect since pairing is still in progress", but the spec
says "The Peripheral shall not send the Security Request command if the
pairing procedure is in progress" (v5.3 Vol. 3 Part H 2.4.6).

Original-Fixed: 80727
Test: `fx test bt-host-sm-tests`, added SMP_Phase3Test.UnexpectedOpCode
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555681
Original-Revision: 6102aaf41608093abdcc01d24687524843b4c849
GitOrigin-RevId: 371417d38055fe3ee89cd58b546d6dcbc3e5a42f
Change-Id: I129d0580b02f9191bea9cd9f5c729d63d7b46152