[roll] Roll fuchsia [starnix][pipes] Implement blocking partial writes
When writing to a pipe in blocking mode, if there is a partial write,
then the operation will continue to block waiting for more space to
become available in the pipe. If the reader shuts down, then the write
completes as a short write an the caller receives a SIGPIPE signal.
This CL updates our behavior to match the description above instead of
simply returning a short write immediately.

After this change, we pass all but a few tests in pipe_test. This CL
switches over to a denylist of tests we do not pass instead of the
previous allowlist of tests we do pass.

The remaining tests either require procfs, statfs, or O_LARGEFILE.

We now pass 114 tests in this test suite.

Test: PipeTest.BlockPartialWriteClosed, ReadFromClosedFd, Streaming
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555753
Original-Revision: 4f7a1d7b628798b8fd09c1b6dd0d1f50cfece654
GitOrigin-RevId: 4057062ce53507f71cf94037e5405e882a0b78f0
Change-Id: Ic9c198657253fc6be6877cfd927b6f6e7a3bfe35