[roll] Roll experiences [ermine] Fix loss of focus when app is visible
This change fixes the issue where moving focus using keyboard arrow and
tab keys gets lost when an app is running. The root cause was that the
app was getting the focus even when overlays were visible.

Fix was to keep the focus within the scope of overlays.

Original-Fixed: 80346

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/556027
Original-Revision: dad7d161732a82356c5f3b2fc6f46aee7380ddf8
GitOrigin-RevId: a6819fe9ae6d47f6fd1fcd8c5177726fac92dc14
Change-Id: I00ea4eda191cce36d75841716250261def5e1608