[roll] Roll infra/recipes [cts] Change CTS artifacts to be relative to build dir
cts_artifacts.json has or will be changed to list paths relative to
the build dir, rather than the checkout dir. This recipe change
accommodates for that change.

This is a hard transition, but CTS is not supported in production yet
so this is okay to land and temporarily break things.

Original-Bug: 79984
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/556021
Original-Revision: 259dddc980a09997d07df38212e4aac9b13a5636
GitOrigin-RevId: aa9f27e0919880f4ccca00f10dedf1f5d8e3657d
Change-Id: I2b16934dd9aea22fc010b1c849a8f88a0e3bdaae