[roll] Roll fuchsia [kernel][virtual_alloc] Add virtual allocator
The virtual_alloc library implements an allocator that can be given a
range of virtual memory and then provides a page granuale allocator
inside that. It has no dependencies on the heap and only uses the pmm
and MMU (which also have no heap dependencies).

This commit just adds the library but leaves it unused.

Original-Bug: 30973
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514687
Original-Revision: abefae34eb603ab0609bf1183365337e5b04b55e
GitOrigin-RevId: fd8f5d4f769a0545c32e7fccbf30233573f1ec41
Change-Id: I6a87cca466443fe5a6a475fdf44a159773002086