[roll] Roll fuchsia [debugger] Fix a memory leak in Curl
Curl keeps a self reference when an async transfer is in progress, which
makes it possible to fire a request and forget it. However, if the
message loop terminates before the transfer finishes, the self reference
will never be freed and causes a memory leak.

Fix this issue by reorganizing the ownership of Curl and
Curl::MultiHandle. In the past the Curl class manages itself and owns
the MultiHandle, while the MultiHandle doesn't interact with the life
cycle of the message loop. In the new design, Curl no longer keeps a
self reference or owns MultiHandle any more. Instead, it will save
itself in the MultiHandle to avoid being destructed when the transfer is
in progress, while the MultiHandle will save itself in the closures sent
to the message loop, so that when the message loop terminates, all
dangling Curls and MultiHandle will be deleted.

The interface of MessageLoop::WatchFD is changed to accept a
fit::function rather than a FDWatcher object to allow saving a closure.

Original-Fixed: 79164
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554766
Original-Revision: 870d505040c5564f466c61b25ad1890bc17e2584
GitOrigin-RevId: 4d6651b3f84ed40a70454fc08945d94e8897d09f
Change-Id: I9ccbd645d235e287e90490305546daf50118349d