[roll] Roll fuchsia [cmpctmalloc] Increase maximum heap growth to 2MiB
This changes the maximum heap growth amount from 1MiB to 2MiB.

The intention of growing by 2MiB, instead of 1MiB, is so that when
using a virtual heap, that requires mapping in new pages, large
pages can be used, saving on MMU memory usage.

In this worst case this could increase memory usage of the heap by
1MiB. However, this requires the heap to have already grown such
that 2MiB is 12% (the heap grow ratio) of the heap at which point
an extra 1MiB is a relatively small portion of memory.

Original-Bug: 30973
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528795
Original-Revision: acad24b86e026420fe70dfc709a21cf3a621516a
GitOrigin-RevId: 6bf3f7d0c22ae6fd8c3e4b01f6211959b71a110c
Change-Id: Ia813c349e2650710cc64d4bc5e408f0ddcb69cfd