[roll] Roll fuchsia [iwlwifi] Don't require a newline in diagnostics
Remove runtime check that triggers an assert when a diagnostic
format string doesn't contain a newline. In fuchsia, zxlogf()
doesn't require a newline (and will in fact generate a blank
line if we use one), so we strip it from diagnostics.

This CL simply removes the code that forced the runtime assert()
when a newline wasn't present.

Original-Bug: 77650
Test: Manually verified that a diagnostic without a newline doesn't
      cause a crash. Also ran a quick smoke test to verify that
      things are generally working as before.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556183
Original-Revision: c8287966a6e5c8d025eaecde8b21d969ce3eba1a
GitOrigin-RevId: ab3c61622a160c433e0fbf7fb05bc4f01c6acb42
Change-Id: Idc70287cc23522ea7a7e5226a8d4013868608b99