[roll] Roll fuchsia [mold] Replaced bit set with a range set.
The old implementation of the set of composition IDs was backed by
a bit set since it was optimized for memory only. This led to
somewhat poor performance (O(N)) when using quite a lot of layers.
The bit set is being replaced in this CL with a range set that is
pretty memory efficient, but finds new layer IDs in constant time.

Testing: fx test mold-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556381
Original-Revision: a594c031928d3be83fa09453b55f300342630c64
GitOrigin-RevId: 7635f19b23c960eec7433f1cc37cad5ba6b0fe99
Change-Id: I5da186959c390a00d8565a550511a31f383b9395