[roll] Roll fuchsia [docs][diagnostics] Make bitfield diagrams more clear.
It is more useful to see the start index of each space than the index of
the following space. This CL changes the numbering to instead show start
index, and we update the description document to be more clear about how
to interpret the new format.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556405
Original-Revision: d5b31f92155270908842a0f338ed3b069d9c8d5c
GitOrigin-RevId: 0b461b44c0367157420f56cc89301ec5ce424015
Change-Id: Ie73bd9620e4e9384347844cd0355fa79b5a82c73