[roll] Roll fuchsia [mold] Replaced DashMap.
The DashMap used in mold was more of a commodity and not strictly
required given the requirements. However, since in high-resolution
environments it was adding even 1ms+ of latency to drawing, it is
being replaced with a Vec that can be used in parallel by taking
advantage of rayon's functionality.

Testing: fx test surpass-tests && fx test mold-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556341
Original-Revision: 7aab4e89e3cf8fec3a32de40dcf6187178f732b0
GitOrigin-RevId: c0cac64782a5d0bb69d9100d40dd03ee36480e96
Change-Id: I6ec3d2570bc1dbddf4040c363919493615487895