[roll] Roll infra/recipes [recipes] Add incremental field to fuchsia.proto
Incremental builds are currently enabled using an experimental flag.
This is not ideal now that we are using incremental builds in
production; the flag should be migrated to a more permanent codepath.
This CL begins that process by adding an incremental flag to the build
spec proto.

Original-Bug: 64127
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/556035
Original-Revision: 43b1f14bb45449663150c515f73d7c7805c9a68a
GitOrigin-RevId: a53988be1b32edfe062eee9d39c3b3094a754520
Change-Id: Icf43e1f2f90a23cdef4edeb47e31c9d3ae1844bc