[roll] Roll fuchsia [host][blobfs] Convert depfile to buffered FILE*
The host blobfs tool reads files, hashes them, and compresses them in a
small collection of threads. Each thread takes a blob to process from a
global queue, processes it, and adds back the hashed and compressed
object to a completed queue.

The tool may optionally output information about each blob to a
'depfile', a standard file with each source blob listed. Each thread
writes to the depfile as it processes blobs, which required some
cross-thread synchronization.

Convert the depfile from a mutex + raw FD to a buffered FILE*.
The FILE*'s fwrite() path only writes to an in-memory buffer, which
may be flushed later; this removes a syscall-under-a-mutex that
every thread would have to pay for and batches writes to the
depfile fd.

As part of fxr/404773, this change and the other parts of that fxr/
reduced runtime of 'blobfs create' by ~2% when packing core.x64
--with-base //bundles/buildbot:core on a 16-vCPU VM on GCE.

Original-Bug: 55297 It would be nice if the Fuchsia build was faster
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405102
Original-Revision: 127b765e0f37ad47fcee5324d874b5dbdb315ea7
GitOrigin-RevId: f6c61a0a253fdda71bd2cd31a5ceb7853274619e
Change-Id: I9b1acf663f9deee58488c8b9961a0c92581fd1cc