[roll] Roll fuchsia [starnix] Pass ExitTest.CloseFds
We now represent zombie tasks with a minimal struct that contains the
information we need rather than a complete Task object. This approach
lets us drop the Arc<FdTable> when the Task exits as opposed to when the
zombie is collected, which is required to pass ExitTest.CloseFds.

After this CL, we pass all the tests in exit_test.cm.

Test: ExitTest.CloseFds
Original-Fixed: 79266
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556113
Original-Revision: 16781bee12128786da0146004c5426855a48f8eb
GitOrigin-RevId: ff262dc59a22965cc6a4a034d8dc178f3382c982
Change-Id: I290a4cc83675ebf3e9515ffd0611d0da5f9dc4d5