[roll] Roll fuchsia [display] Rename primary/vc nodes as client
Renamed the `primary` and `vc` nodes to the more generic term
`client` and added a "primary" boolean property to convey that
semantic meaning. The more generic term allows a selector such as
"client-*" to filter for all clients, at the expense of the ability
to specifically filter for primary or virtcon.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555488
Original-Revision: 336b3e77046ca061123fca9f2d6e932a9ee54d01
GitOrigin-RevId: 8fb048c15a8eb5ece200b30b8e4e310969d8c778
Change-Id: I7d601394db41681889839a9b7fba5f1ca6061869