[roll] Roll fuchsia [zircon] remove old comment
This CL removes a comment in the arch_context_switch function. The comment was referring as to why the function had the `__attribute__((target("fsgsbase")))` annotation. That annotation was removed in CL 409874[0], where the kernel is now being compiled with `-mfsgsbase`.

https://fuchsia-review.googlesource.com/c/fuchsia/+/409874

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556521
Original-Revision: cb02437f52b8065b4e2cce321602a67398ddeb3a
GitOrigin-RevId: 91620a8f1698e1144458dcd40d4f2ca2ca354da6
Change-Id: I7856b052d4b0f5938017c7e50781d41fbe284a3a