[roll] Roll fuchsia [size_checker] Check the Update Package size when absent from blobfs
The Update Package is being removed from blobfs, but we need to continue
verifying that the size is within budget. We are adding a separate
section to the size_checker JSON that allows us to verify the size of
blobs even if they are not assembled into BlobFS.

This verification is handled through use of the blobfs-compression tool,
which does not de-duplicate blobs, and sometimes over-estimates the
size. This is alright for the Update Package, because the amount saved
by de-duplication is negligible (~9kib).

Original-Bug: 58645
Test: fx test size_checker_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554320
Original-Revision: 892f55b89a9024766d6a068fa50e9ea680c06fa8
GitOrigin-RevId: 285fe69f745cde340db6d618d8b4c8062f592f12
Change-Id: I90a9d626d01317b71d53722cc5a3f5b00daea24e