[roll] Roll fuchsia [bt][mock-piconet-server] Fix bug in generating ServiceFoundResponse
The ServiceFoundResponse for protocols that specified RFCOMM should
include the RFCOMM channel number so that peers in the piconet can
correctly connect to the service. This was incorrectly removed in
fxrev.dev/553171 in which all RFCOMM-related parsing was removed.

Update the local ServiceRecord type to store the primary protocol
directly. This simplifies building the ServiceFound response and
prevents any sort of information loss in the conversion process.

Update test helpers to validate `peer_id` in ServiceFound responses.

Original-Bug: 79742
Test: fx test mock-piconet-server-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555943
Original-Revision: e360493b7588342f2fde293d8cadeda256a6c503
GitOrigin-RevId: 1c89e8aa87d57c26443dbc44618fb684b12a2f0d
Change-Id: Ie9f7666a21db074cc7460acf0e242dd1bbed07dc