[roll] Roll fuchsia [carnelian] Use BTreeMap for compositions.
This makes it possible to change the number of layers in a
composition while maintaining a stable order. Critical for
good text grid performance.

This reverses the order of some scene building code to
become top to bottom to match the btreemap.

The textgrid example has been updated to take advantage
of this and supports partial updates as a result.

Test: carnelian-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556321
Original-Revision: 4b2706745576b26613958303bc28770eb27fc4dc
GitOrigin-RevId: 81805e03a29827c2fbdb2764d7e2be81bc884c12
Change-Id: Ic087475ba44cf862ce5ead38a5a23d42cd3de167