[roll] Roll fuchsia [build] Integrate verify_routes into the Fuchsia build.
This change integrates the ffx scrutiny verify routes command into the
build so that the build will fail if a use statement from a protocol or
directory is introduced without a matching expose-offer chain in the
build. This is to prevent invalid routes from being added to the image.

This does not include components that are not linked up at some point to
root.cml and only verifies v2 components.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545001
Original-Revision: ecf548552c9031c94d44f1433681260f8cd8560e
GitOrigin-RevId: e570b05a425b1cc78d17623f4bcae722f052d44b
Change-Id: I21982bfe768074fdb6356134b37a063b68807ccc