[roll] Roll fuchsia [build][go] Ensure the right version of 'ar' is used for C archives
It was discovered by running our build in a container with a minimal
filesystem image that the "go build" command will invoke the host
'ar' tool when building a c-archive library. This failed because
there was no 'ar' in the PATH in the container environment.

This change ensures that the correct toolchain-specific 'ar' is
used instead in this case.

Original-Bug: None

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552183
Original-Revision: 45a85bbeab39be98abf7aec98bf130a5995fbcd8
GitOrigin-RevId: 94f1a5d3f8871d2b7a6c726e1b3ec3afb65fd4de
Change-Id: Iea20f51455c774004909782b8b3720d4240aba39