[roll] Roll fuchsia [netstack] Log when interface watcher queue full
Log a warning when a `fuchsia.net.interfaces/Watcher` is closed
because the event queue is full. This may be because the client
is no longer interested in the events and have stopped calling
`Watcher.Watch`, in which case it should have closed the channel;
the server is generating spurious events or unexpectedly introducing
latency while handling the requests; or some environmental factor
such as unexpected latency anywhere in the FIDL method call chain.

Also lower the maximum queue length to 128 from 512 (which was
excessively conservative).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557050
Original-Revision: f26d3c85e1bd62c6ce48c0af276c87a0b239f9b6
GitOrigin-RevId: 38ccab2c5f3a798c5e7c38ef1d36aae8247d0620
Change-Id: I7b5f4e0aa4dec88e678dc7faafc57d978f17fdf8