[roll] Roll fuchsia [kernel][crashlog] Add "persistent" dlog support
Add the ability to share some of our Very Special Crashlog memory in
order to produce a persistent dlog.

The persistent dlog is a ring buffer into which we always store the
most recent messages which get written to the dlog. Data written are
flushed from cache to RAM each time, allowing them to be recovered and
added to a crashlog in the case of a spontaneous reboot. Writes will
never block waiting for space, or waiting for a UART. Older data are
simply overwritten.

Tests: new kernel unit tests
Original-Bug: 70724
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506930
Original-Revision: b04e6240a84be46368f5e21575a3982eb9b30470
GitOrigin-RevId: fa0f956a556943e8d2bc61d77aade4ce2a1629dc
Change-Id: Ic79b8bd9f8d1017ed57ed2de7754b3e0a2439643