[roll] Roll fuchsia [kernel][arm64] Ensure EL2 page tables are disabled during mexec.
We disable caches and the MMU for EL1, but fail to do so for EL2.

If a guest happened to be running when we call mexec, we will attempt to
jump into the new kernel with the MMU and caches enabled, violating the
assumptions of the new kernel (and possibly writing over the page tables
while we shuffle data around for the new kernel).

Update mexec to also disable the MMU and caches for EL2 if we're using
it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550927
Original-Revision: 442c8ee4e1f2009f9babf06d64a26bfa965f45e3
GitOrigin-RevId: 3cbd24f89b68d719d927f47cc5b3f191462a5aa1
Change-Id: Icea4a29d7141bfff1c3f7624ca3fda5750465e7d