[roll] Roll fuchsia [memalloc] Introduce Pool::UpdateFreeRamSubranges()
This method converts all kFreeRam subranges of any given range to a
given, extended type.

The utility of this operation lies in situations where there is a
special range that we ultimately want reserved for "something" later on,
but it is immaterial what occupies it in the meantime, so long as
nothing is further allocated from there. This will be required in the
case of loading a fixed-address kernel image, where we would want to
prevent page tables - which must persist across the boot - from being
allocated out of the that load range.

Original-Bug: 77359
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545948
Original-Revision: 49c27c40c50cfb7dcffbc8ec4aaea708e368e2b8
GitOrigin-RevId: 06e71379f05c3c28dc40000ef54696ca372419b5
Change-Id: Ic62c0bfaffde2a30f27033c27ea8c1ced5183721