[roll] Roll fuchsia [component_manager] Implement reboot-on-terminate
This change implements the 'on_terminate: reboot' feature for parity
with v1 critical_components, as described in the linked bug.

This introduces some changes to the model to implement the feature. The
top instance (component_manager's instance) now has a state object to
hold a reference to its child (i.e., the root) and the reboot task, if
in progress. The child pointer allows the top instance to access the
statecontrol.Admin protocol exposed from the root.

This change does not enable the feature, so clients may not use it yet.

Original-Bug: 80365
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553607
Original-Revision: b56cb4a8a904c9a6379488a0d14cc1e3337744af
GitOrigin-RevId: 8890f63db70f5027a8fac2ce9e873b5bfdc127d5
Change-Id: Id0abb62f6a7e14acba0a3b14f95d7cc0e1854b68