[roll] Roll fuchsia [lazy-init] Remove address-of operator.
Remove the address-of operator from LazyInit<>, replacing it with a
specific method (GetAddressUnchecked) instead.  The primary purpose of
overloading it in the first place was to allow code to take the
address of the underlying T at compile time to do things like
initialize a global array of pointers.  The runtime checks present in
the Get() method would prevent this in some cases.

Instead, users may now use GetAddressUnchecked to explicitly fetch the
address of the underlying storage without hitting the runtime checks.

Also, fix a number of places where the code was inappropriately using
the & operator to fetch the address of the underlying storage, and
accidentally bypassing any checks in the process.

Test: existing unit tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557801
Original-Revision: d6b1c5bb5726ccb883a38550b6ad6f0213e71b47
GitOrigin-RevId: a3d162e1a5f9c39ae4e59ccea48ac97e29fdd03e
Change-Id: I9239ce30211be6cd7390ab53a76fd08de6d16010