[roll] Roll fuchsia [kernel][io] Refactor Linebuffer helper.
Move the Linebuffer helper class out of lib/io and into the Thread
class as an inner struct.  This allows the storage details of the
linebuffer to be a property of threads while leaving the printing
behavior as a private detail of the io library.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555744
Original-Revision: cdea979bdeeac49cf54438948c991ffef0770c14
GitOrigin-RevId: 0b93cc2dc84dbf212b3d149321fe99820f80918e
Change-Id: I1e045d08237870feb6206caea422c9fe9540a005