[roll] Roll fuchsia [docs][fidl] Fix links in fidl documentation
Fix how links to source code and live documentation
are written in comments in fidl files so that they
render and navigate correctly both locally and in
fuchsia.dev.

All the links fixed in this CL are currently broken
when browsing the corresponding
https://fuchsia.dev/reference/fidl pages.

There is currently no standard for links to docs in
fidl and code comments regarding being absolute or
relative. Currently, they need to be host and path
absolute in order to render correctly in fuchsia.dev.

This CL is not an attempt to define such a standard,
and a definitive decision should result from an
ongoing discussion in Fuchsia Devrel and Tech
Writers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557724
Original-Revision: f3374b214f80edf14bf10e49f97c4142bfb25f01
GitOrigin-RevId: 5a6b2cb218944c455da3ef9fb406d5e949b694c5
Change-Id: I2501f3d10b984dfc39f3a4342427fd075f3f076f