[roll] Roll fuchsia [kernel][hypervisor][arm64] Flush caches on guest set/way operations.
Trap guest set/way cache maintenance operations and use them to (i)
perform full cache cleans on behalf of the guest; and (ii) re-enable MMU
detection logic.

Additionally, stop using HCR_EL2.DC ("direct cache") mode, which can
only be used in limited circumstances, but hides caching bugs in the
circumstances it can't be used.

Background:

When the hypervisor boots an ARM64 guest, it starts the guest with the
HCR_EL2.DC bit set, converting uncached memory accesses into cached
memory accesses.

The DC bit makes the CPU ignore guest requests for the MMU to be
enabled, however, so the hypervisor can't keep it on forever. Instead,
the hypervisor monitors guest accesses to MMU-related system registers
so that once the guest enables the MMU, it can clean caches and clear
the DC bit.

Monitoring guest accesses to MMU-related system registers is expensive:
we need to trap into the hypervisor for each access, and a context
switch has several such accesses. So once the MMU and caches are turned
on in the guest, the hypervisor assumes that the guest is responsible
for managing caches, and stops monitoring.

The problem is, if the guest ever _disables_ the MMU again, the
hypervisor doesn't notice, it doesn't sync caches, and it doesn't set
the HCR_EL2.DC bit again. This leads to random corruption of memory due
to caches/main memory not being synced.

The guest turning off the MMU once it was on didn't tend to happen
previously, but upcoming changes in the Zircon boot path (physboot) need
to do this, hence the need for this change.

Original-Bug: 80567
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557401
Original-Revision: f72af4d145e0a9d10622b29740e8230ad370634a
GitOrigin-RevId: 2442184b515c1189e98e6d0241a0b9bebbff7174
Change-Id: I0b20cc95187636f7d66c83799a2a247b825449c5