[roll] Roll third_party/go pprof: disable flaky inlining test
This test has been flaking on arm64 and amd64 builders for quite a
while. Reproducing is possible, but takes time, and isn't a priority in
the team.  Skipping this test as it also isn't blocking anybody's work
or causing issues.

Original-Bug: 72646
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/557984
Original-Revision: 16f4fc1216643d7bc50add8bf46230501725c0dc
GitOrigin-RevId: 41049ad5bacfdb40c36c0f61307cc3484593fd09
Change-Id: I07946e105eae6ad8ec8583fb0148eca19b1e55b4