[roll] Roll fuchsia [diagnostics] Create diagnostics-kcounter that starts on boot
This is a re-implementation of src/developer/kcounter_inspect in Rust
using the unified diagnostics launcher. It publishes the kcounter
Inspect VMO from kernel as lazy values at the root of the component's
Inspect tree, and updates the VMO whenever it is accessed.

Test:
* A unit test verifies that we call {get,update}_inspect_vmo correctly
  and lazily.
* An Inspect integration test is added to verify the existence of
  handles info in the Inspect tree produced by this new component,
  similar to the integration test for kcounter_inspect.

Start-on-boot is achieved via an empty control protocol that's consumed
by archivist, similar to diagnostics-persistence.

Original-Bug: 79123
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555362
Original-Revision: 4f659a626affcc7957cfa43150d243f5c0041777
GitOrigin-RevId: 428caf6cf47c9e3c797861105e49c7d743b7c00f
Change-Id: I15fb18943b7cf8be8361d9bb193ff85455723591