[roll] Roll fuchsia [fxfs] Fix race between unlink and open
We need to take an open count whilst we hold locks during lookup since
otherwise it's possible for a file to be unlinked and then purged after
lookup but before the open count has been increased.

The other issue was with purging from the node cache. We need to check
that the node we are removing from the cache is the one we expect.

Original-Fixed: 80975
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558007
Original-Revision: 0db39954a74204c2fdeac0f4894d62b56c532fb5
GitOrigin-RevId: 76b02fa39807de639b4724dae67b734ef9d0986b
Change-Id: I9089f7158ae80be5925aabbdbc94471b8a9da39d