[roll] Roll fuchsia [fvdl] Replace use of "curl" with hyper's "get"
In fxr/537543, a comment was made to replace a std::process::Command invoking
the "curl" executable with a call to hyper's "get" command. This change was made,
but a bad commit/upload overwrote it, and so the "curl" command was pushed to
HEAD.

This CL redoes the change from "curl" to "get".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559022
Original-Revision: c2d28704e655e6134bee210843eae89ae6b742b1
GitOrigin-RevId: b708b18a30c7275d73a87161668377c104b20c18
Change-Id: I3ca3db65aee73758ff13bd1c6126c592c610e2f9