[roll] Roll fuchsia [debugger] Correct PC for hardcoded breakpoints
Previously, debug_agent will set PC to the next instruction after the
breakpoint instruction for hardcoded breakpoints, e.g. in abort(). This
creates confusion for users and the unwinder. Change this behavior so
that the adjustment of PC is done on resuming from the exception.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558401
Original-Revision: 8429bd1cce2aa4515917b4845e0645fc16bf369e
GitOrigin-RevId: 728e9edd8472343751a1c00d28d9fcfa9cb46d24
Change-Id: I64f14012762caa4d2eb3305559f7ea2be499f981