[roll] Roll fuchsia [astro][touch][gt92xx] Fix interrupt GPIO resistor
Change the GT92xx interrupt resistor to a pull-up instead of a
pull-down. There is a chance that this will fix the continuous errant
interrupts that happen in rare cases, but should be done anyway in order
to match the original driver.

Original-Bug: 79479
Test: fx test gt92xx-test
Test: Used print-input-report to read touch events on Astro
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559449
Original-Revision: 0fc7ae14d9adb92269ebe269ef6514aa808363e7
GitOrigin-RevId: ffec485fb12695568ea9ca9c79b4fef38585def8
Change-Id: I9b92bdcbbb042e2fb739c45128615100323935dd