[roll] Roll fuchsia [wlan] Add BssTypeSelector and an UNKNOWN value to BssType
The new BssTypeSelector type substitutes the use of BssType
in *ScanRequest struct types which are intended to specify an
acceptable combination of BssType scan results to return.
At present, both softmac and fullmac drivers effectively ignore the
bss_type_selector argument, so this change is only semantically
meaningful.

The new UNKNOWN value in BssType is used for BssDescription which is the
value associated with each scan result. Since the BSS type is not
guaranteed to be readable or recognizable, the BssType now has room
for an UNKNOWN value.

Original-Bug: 79613
Test: fx test //src/connectivity/wlan

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551584
Original-Revision: c3490e77424119228d5bce05eb96e369b2859dbf
GitOrigin-RevId: 7ad6be0239b79a1581414738b8a1aaa0f627b811
Change-Id: I959cd1e68801bbb96d0afb60417e68078be279b5