[roll] Roll fuchsia [i2c][amlogic] Don't log errors for NACKs
According to the I2C spec there are valid reasons for devices to NACK
bytes from the host. For example, the SHTV3 sensor on Nelson NACKs
requests to read temperature data when a measurement is in progress.
Remove the error message in the Amlogic I2C driver to eliminate log spam
in these cases, and instead rely on client drivers to log when
unexpected errors occur.

Original-Fixed: 81383
Test: No more I2C error messages on Nelson
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560367
Original-Revision: ca4ed54243785bea38841ccaf5f590eab767c935
GitOrigin-RevId: 96a83b683f21d28f35b16c20dd8c6d704394603f
Change-Id: Ie4ccb225965fee16aff3e2caa16245a1f90b86d0