[roll] Roll fuchsia [bt][hci] ensure AdvertisingHandleMap issues AdvertisingHandles in sequential order
There is no guarantee of order or sequence in the values that the
AdvertisingHandleMap returns except that it returns a valid handle in
the range of valid handles and that the handle is currently unused.
However, for the internal consistency of the class and keeping it
clearer to read for future authors, ensure that it returns
AdvertisingHandles in sequential order.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559861
Original-Revision: 19b352f32b807b61dfed61fce8d8132090255bf8
GitOrigin-RevId: 5e733e7fbd2b1efc8b78cb35a1308f390bbc9186
Change-Id: I8b7bd0c32ceb6e2bdfb3dc937f37ebe270b9914f