[roll] Roll fuchsia [fidl-lint] Remove repeater checks
Repeater checks attempt to enforce the aspect of the FIDL style guide
that says to avoid repeating a component from the parent context's
name (unless it serves to disambiguate).

This check has never been enabled due to false positives - given the
amount of time this feature has existed but not been touched,
we opt to remove this feature rather than continue maintaining it
(for example, updating it to support anonymous layouts). This is
done by simplifying the context stack from a deque<Context> into
a stack<string> representing the layout kinds (the deque -> stack
is an independent change to improve readability).

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler
--gtest_filter='LintFindingsTests.*'

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560443
Original-Revision: 7db888a081ced168c3e0df71a012e2b6ac7df634
GitOrigin-RevId: 68e75dada25fede9b639964518b44044b4908444
Change-Id: I632707f9c9ec00c54fcbae642157334c2f7cac38