[roll] Roll fuchsia [fxfs] Fix race in graveyard reaper
When the reaper task finished it would remove its own future, but
that would mean that if wait_for_reap was called just at that moment,
it wouldn't wait, and would just continue.  The problem is that at
that point the Arc reference hadn't been dropped, but test code
depends upon this. The solution is to not remove the future when the
task finishes.

Also, the paused state wasn't doing anything so I got rid of that and we
can just use an Option instead.

Original-Fixed: 80777
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559902
Original-Revision: 6bb16c0c91a08d299f6fcbe87132c46a2177ed3a
GitOrigin-RevId: 155983e0b4e330fbe575bd0fa9643f6d5b1a0ffb
Change-Id: Ic73eef14a9cd2898087ea47a1eb632da14f937cc