[roll] Roll fuchsia [blobfs] Update blob command-line flags
Changes the blobfs tools from the flag
  --blob_layout_format=[padded|compact]
to
  --deprecated_padded_format
when the old format is needed, and nothing when it's not. This is to
help document that this format will not be supported forever, and to
encourage the compact format by making it the default.

Both flags are currently supported in zircon/tools/blobfs because there
are some out-of-tree scripts that use the existing flag.

There should be no behavior change as all builds should get the same
flag they had before, only the way it's specified has changed.

Original-Bug: 79775
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559191
Original-Revision: 15765e8de60d4286ba2a4cdadbd0d99fad812986
GitOrigin-RevId: 26d5c46d26076bab8a832a97e7f6ffba323563ea
Change-Id: Ic2f27cea322ca5cded56162f82fbe50aafca5aa3