[roll] Roll fuchsia [assembly] Allow clients to specify the output ZBI name
The board config for the Image Assembler now uses board.zbi.name to
allow clients to specify the output ZBI name. "fuchsia" for
"fuchsia.zbi". A "signed" suffix is appended if a ZBI signing tool is
specified in the board config.

The Image Assembler generates two ZBIs:
1) <board.zbi.name>.zbi
2) <board.zbi.name>.zbi.signed (if signing tool is specified)

The name of the ZBI as mentioned in the Update Package is:
1) zbi
 OR
2) zbi.signed (if signing tool is specified)

Original-Bug: 79886
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559886
Original-Revision: 98e754f5ce500dec03fb4bb32e585c5b4db12301
GitOrigin-RevId: 571367220a90d427d8b8646a67a3faf2c5dd872b
Change-Id: Id780a231bb87b9dff5fda5a5b328ee9525c671ef