[roll] Roll fuchsia [fuchsia-async] Fix deadlock in emulated channel
The emulated zircon handle implementation shards the process global
handle table to reduce lock contention.  It also needs to be careful to
not re-entrantly lock the same shard, or that operation will deadlock.
A test flake revealed a rare edge case which may end up in this
situation.

When sending a channel through a closed channel, the write_etc call will
detect that the outer channel is closed and close all handles that were
given to the write_etc call, however, it was doing that with the outer
channel's shard locked.  Closing a handle also locks the shard for that
handle, so if both channels were in the same shard, that write would
deadlock.

This change modifies the error path to defer closing the handles until
after it has released the shard's lock, avoiding the deadlock.

Looking through other APIs here, this seems to be the only situation
where this kind of deadlock could occur.

Original-Fixed: 81561
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561130
Original-Revision: 416a32567236c91636200770a28b12b3ffa27380
GitOrigin-RevId: 665cb713facbab31319f1c2757873b5143b5440d
Change-Id: If196c262ca4c48534220c62ca697fac763d7a862