[roll] Roll fuchsia [blobfs] Remove old pager ifdefs.
Now that the new pager has been enabled for some time, we can remove the
old pager code. This patch removes the ifdefs and temporary names
associated with the transition. It does not remove any of the old pager
machinery in blobfs/pager which will be refactored in a followup.

Original-Bug: 51111
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560964
Original-Revision: 47ee25a54aa0ef0c376e76e820216507613582fb
GitOrigin-RevId: 8362755b751ac815634cd25f3f89b92fa1eddb47
Change-Id: I0dfc06125268837ef3375486bd58cf09ad6048eb