[roll] Roll fuchsia [netstack] Generalize DHCP integration tests
The utility method currently used by the DHCP integration tests couples
concepts in the netemul realm with concepts in the DHCP realm. This
coupling significantly limits the topologies we can test.

This change replaces that util with one which accepts netemul
configuration and netstack/dhcp configuration independently. This allows
the callers to mix-and-match network and DHCP topologies as desired. A
side effect of this refactor is that the test body references data
allocated on the stack within the setup code. In order to support this
pattern, we moved away from the `test_case` macro (which destroys stack
allocated variables before invoking the test body).

We also removed a multi-network variant after concluding that it was
just testing non-interference across broadcast domains and therefore
didn't add anything useful for DHCP. Future work will extend the usage
of the new util to more exotic configurations.

Multiply: netstack_dhcp_integration_test
Original-Bug: 73252
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520300
Original-Revision: 244f1b80666803f8f2dc80f4370a077b3c77e109
GitOrigin-RevId: 734840d75e43deaec9e876059f63648c0063f5af
Change-Id: I15b698354687213ccc6b0d27c124f1d229e402a6