[roll] Roll fuchsia [bt][gap] Be resilient to duplicate bonds in store
In some system configuration flows, it is possible for the same
underlying bond (i.e. security information associated with a local
host address - peer address tuple) to be associated with multiple Peer
Ids, and thus added to the Store component multiple times. This can
cause issues, as bt-host fails to restore a bond if it contains an
address already associated with a different PeerId.

To be resilient to this, we check for duplicate local address-peer
address combinations when loading bonds from the store during Stash
initialization. If the underlying security information is the same, we
pick one of the bonds to store in memory and delete the duplicates from
the store. If any of the duplicate address combinations contain
distinct underlying bonds, we fail the stash initialization process, as
this indicates an invalid state in the store.

We choose to handle this at bond-loading time instead of bond-storage
time because it is not as simple to distinguish between the allowed
"replace an old bond associated with this address" flow from "add two
distinct bonds for this address" during bond storage.

Original-Bug: b/193454286
Test: Added `new_stash_filters_duplicate_bonds` and `new_stash_fails_
      loading_same_addrs_different_bond` to `bt-gap-unittests`.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560662
Original-Revision: 536b693d0be4854d06886337a0ece45294e3b1c3
GitOrigin-RevId: 59a3746f54969060f5c713bf922fc65826c63a92
Change-Id: I9bde831210d8aacbed54665405654a44377e2aa4