[roll] Roll fuchsia [build][golden_go_proto] Ignore compiler/plugin versions.
This ensures that the compiler and plugin versions, embedded inside
the generated .pb.go files, are ignored when comparing them with
the golden/reference file.

This ensures that a real difference is an error, instead of a warning,
and that rolling a new version of the compiler or plugin is not going
to break things needlessly (which was the reason for the warning in
the first place).

Original-Bug: 72810
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561001
Original-Revision: b053d162c2085aa3983eecf56ff9e6ab8491867c
GitOrigin-RevId: 7307c11958746f74234261650843fb644313c454
Change-Id: I31032c40208d6f4f983076d499e5614dc3806d51