[roll] Roll fuchsia [kernel][x86] Fix MWAIT idle loop preemption bugs
The Fuchsia MWAIT idle loop is run from thread context,
on a dedicated per-CPU idle thread. Whenever the thread
is scheduled (when no other thread is runnable), it
attempts to select an idle state to enter and executes
the MWAIT instruction to idle the CPU. Wakeups are
handled by stores to a per-cpu variable, watched by
MWAIT.

The idle loop has a 'preamble' before executing MWAIT
and a 'postamble' after waking up from MWAIT; in the
latter, we record how long the idle time was and we
conditionally execute a security mitigation for
Spectre-RSB.

Unfortunately, preemption was not disabled while the
idle loop was selecting an idle state / MWAITing;
because of that, a local timer expiring or other
rescheduling event would cause us to switch threads
without executing the postamble; this would cause us to
both misrecord idle time and to potentially skip the
Spectre-RSB mitigations.

Fix this by disabling preemption around MWAIT and by
explicitly testing the preempt pending bit.

Original-Bug: 33667 Spectre mitigations?

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540103
Original-Revision: 39a0acf6eac1c8849eec2572c859525ab115734c
GitOrigin-RevId: f4d73075d19182acfe28267635ec544ca833a47a
Change-Id: I4980db1a2bc136b6be37e74f71f0f65094f52a65