[roll] Roll fuchsia [sshutil] Consider EOF to be a ConnectionError
An EOF when trying to run an SSH command likely indicates that the SSH
client was closed due to a failed keep-alive ping. This should be
classified as a connection error so that testrunner will retry the
command instead of classifying it as a test failure in case an SSH flake
manifests in this way.

Original-Fixed: 81609
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561721
Original-Revision: b701b2d35a4fdf6353fbb48db07b46d8ed852759
GitOrigin-RevId: c39e665526d6d51ba30676239507d0be29c4cc01
Change-Id: I2f970cb1bcb2ca5411c2fac0faef1bf1e4acf0e2