[roll] Roll fuchsia [kernel]: Remove serial acpi handlers.
As of recently, the boot shims parse the ACPI table descriptors and
synthesize a serial zbi item, which contains the configurations from the
uart in the acpi table.

Also as of recently the command line is pulled from BootOptions that
already handles this cases, in which it ends up using the zbi item for
its serial configuration

This CL just removes dead code, as a result of the mentioned changes.

Original-Bug: 80903 # Fixed
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561422
Original-Revision: 18687d43d7e3ff85306461fe97317bead240b51f
GitOrigin-RevId: 014c4ade5250d341fc001a55099224773a287b1a
Change-Id: I9241185a8fb2064e467d01a22543e6b0e22d15aa