[roll] Roll fuchsia [fidl][summary] Start version tracking summaries
This commit adds the API summary files [1] to each FIDL library.
The idea is that from a certain point onwards, we keep these files in
sync with the contents of the library.

The presence of these files is not yet enforced.  Enforcement will be
turned on once all FIDL files which need this treatment have been
equipped with the summary files.  Enforcement is similar to now `.api`
files are handled today.

[1]: https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0076_fidl_api_summaries?hl=en

Original-Bug: 68951
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537862
Original-Revision: 759de0da630ed07c68a1f3a64ab1c0c73f0cc550
GitOrigin-RevId: 7cf39fa86e069dd3cc09a58d5fd477dc50a0e84c
Change-Id: Id6f8986d4aa79c54cd842d77ff5d37c1fe3998d1