[roll] Roll fuchsia [carnelian] Support single buffer rendering
Add a configuration option to set the number of buffers to use
when rendering to the display controller. Handle the case
where that number is 1.

This is in preparation for supporting display controllers that don’t
support multiple buffers.

Original-Bug: 79511

Testing: manual
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561603
Original-Revision: 4baa91db4775cc0ba94fb6375c046e428dde6301
GitOrigin-RevId: b61712cb83c035a05c83276181a2ebf0cbe63551
Change-Id: I753a5d8d9b4a74fc304aa08d6d033321b1653a71