[roll] Roll experiences [ermine] Fix view hittest after app switcher selects it
This change fixes the issue where the view selected using the app
switcher, does not respond to pointer clicks. This is because the app
switcher is an overlay and we disable hittest when the switcher is
visible. But it is not enabled correctly when the switcher is dismissed.
It needs to be dismissed when the target view has received focus.

Also fixes an exception thrown in channel_service, where we were reusing
an already bound proxy reference.

Original-Fixed: 81594,81358
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/561981
Original-Revision: c4c2887a3d06f5a6886425d239b0ef11daddfff5
GitOrigin-RevId: 777957325025150a17148bd6df8d7c8b010629d8
Change-Id: I1269c46e87cd3a1d0cd0b2ef4728951c6f5940b3