[roll] Roll fuchsia [kernel] Remove spurious error logging in zx_system_mexec_get_payload()
zx_system_mexec_payload_get() is a bit funny in that it writes to a
user-provided buffer, but the user has no idea how large to make that
buffer. The workaround to that is that the syscall will return
ZX_ERR_BUFFER_TOO_SMALL in that case, and the user would react by making
the syscall again with a larger buffer. This change removes error
printing in the branch that returns ZX_ERR_BUFFER_TOO_SMALL; it's
spurious given the above.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561507
Original-Revision: f2324a4747e4923a4508e7625faa35d57aa8f136
GitOrigin-RevId: 03d44c151717a1b6bd5c2552100208b546b2fff2
Change-Id: I42f4ba83f76ebc3af7952c864d347ab785658cc9