[roll] Roll fuchsia [blobfs] Move merkel mapping to blob verifier.
Before, the Blob kept the fzl::OwnedVmoMapper associated with the Merkle
tree. A pointer to this was given to the BlobVerifier, but this is
stored very far away such that the relationship of the ownership is not
clear. I believe during cleanup the verifier actually has an invalid
pointer, but we're OK because we don't try to verify anything at this
point.

This patch moves ownership of the mapping to the BlobVerifier so it can
be closer to the pointer that references it. By passing the BlobLayout
in to the verifier, the API could also be simplified in various ways.

Original-Bug: 79611
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561890
Original-Revision: f93e383320d34b2de419dfe72083471d7575dd24
GitOrigin-RevId: 9c35d390b6dbee68ea9a6e44310ba6a923d67292
Change-Id: Ic59be57b7c93c6b881aa9404d13d2fd8fd48da24