[roll] Roll fuchsia [fidl][dart] Validate V2 wire format inlining in dart
Values can be stored inline or out of line in v2 envelopes, but there is
a canonical representation based on the value size. This CL adds
validation that the canonical representation is used when decoding in
dart.

Original-Bug: 80880
Original-Bug: 81180
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561688
Original-Revision: 29c728fd7d513c4e5f2211fb10639e53c281cee5
GitOrigin-RevId: edfb2c889aceb0708c37b5092645c8712667f3c7
Change-Id: I34b2a26ab0f39d70b2f2a378a3e6b26b28f58f16