[roll] Roll fuchsia [fidl][transformer] More V2 GIDL tests & fix transformer bug
This CL adds a variety of V2 GIDL tests to the existing ones.
There are still a number that still need to be converted.

When enabling these tests, a bug in the transformer was
discovered. Namely that it in some cases it was possible to
read and write past the buffer size.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561662
Original-Revision: c6b32e64227fab111230d926a1b208b2e24dccc6
GitOrigin-RevId: b8c6eb51c63f89053803e3e81286a1f271cf2417
Change-Id: I75a1a4e9919c756e858419be8995341aa5f1898f