[roll] Roll third_party/go [fidl][go] Fix header flag byte for v2 wire format
On send side, the wrong flag byte was used. This fixes the bug and
switches to the last flag byte.

This was detected when running the compatibility tests. No test is
included in this CL because it will be tested once compatibility tests
are enabled and I want to fix the issue before then so it isn't
forgotten.

Original-Bug: 80348
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/562067
Original-Revision: 11cc1ae01cbcb6bf38121aa564bd09eac5002eae
GitOrigin-RevId: b7297042376ba5c2bdfc9e804e437923f65a313e
Change-Id: I5db01ebcf7c96ee6a3da6f1d2915fc14406c2613