[roll] Roll fuchsia [power] Use new connect_to_driver implementation
Replaces the crate::utils::connect_to_driver function with a similar
implementation from elsewhere in the codebase. The new implementation
was recommended by the FDF team while investigating a bug in connecting
to a driver's topological path. The bug would result in the Power
Manager attempting to connect to a driver before it was fully ready to
receive requests (likely related to improper directory file watching).

The new implementation also has the benefit of using async Rust crates
for file watching. This is a clear improvement and is preferred over the
existing implementation's usage of blocking FDIO calls.

Test: fx test power-manager-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558750
Original-Revision: ed27dc2457545e2628faa2aeddae48c687118849
GitOrigin-RevId: 21e84f1814b4fe7b0bd0fd2e6b7a48876fb83605
Change-Id: If1e758d43b68a14897271711e25115d65c87bb57