[roll] Roll fuchsia [build][go] Hardlink instead of symlinking source files
This makes it possible to use `//go:embed` directives in Go code as long
as the embedded file is declared in the `sources` of the GN target that
builds the Go file. `//go:embed` doesn't support symlinks.

This is a bit of a hacky workaround; a better long-term solution is
described in the linked bug.

Original-Bug: 81748
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562221
Original-Revision: b6c122bedfc4abb3683870c9d85d53d3271303fd
GitOrigin-RevId: 09f67313a37381e60a04c0eb81d070a9776bfe0d
Change-Id: I8f5f47b8116cde24cc6444fc61eb4a9234e778c3