[roll] Roll infra/recipes [subbuild] Inherit "led_cas_recipe_bundle" property
If a led job is launched using `led edit-recipe-bundle -property-only`,
it's up to recipe_bootstrap to download and execute the recipe bundle
referenced by the `led_cas_recipe_bundle` property. To make sure the
subbuild uses the same recipe bundle, it must inherit that property from
the parent build.

Note that we still unconditionally call
`api.led.inject_input_recipes()`, which causes the subbuild to inherit
the CAS inputs of the parent. That is safe because a led job launched
using `led edit-recipe-bundle -property-only` will generally not have
any associated CAS inputs, so the executable won't be overridden.

The only time that the parent might have associated CAS inputs is if
it's being used to test a locally built version of recipe_bootstrap, in
which case it's desirable for the subbuild to use the same version of
recipe_bootstrap *and* the same recipe bundle as the parent (unless,
like for Mac builders, the subbuild uses a different OS/arch than the
parent... but it's unlikely we'll care about testing a locally built
recipe_bootstrap specifically on a Mac builder).

So we still want to `inject_input_recipes()`, even though it's a
misnomer in this case since the CAS inputs contain a recipe_bootstrap
executable rather than a recipe bundle.

Original-Bug: 81568
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/562222
Original-Revision: d0758c597f7a2101fba0e616e7028111332ce21a
GitOrigin-RevId: 168f6c4565508a01fa17890a6f3ae2e1da90a44d
Change-Id: I420626b4edf1e9ce12e431aea586c56b3b9b4088