[roll] Roll fuchsia [wlan][iwlwifi] Remove unnecessary assertion for associaiton.
The original code (for 7265) has incorrect assumption that
the driver cannot support the chipsets with low latency mode.
It turns out the assumption is overthinking.

We are still able to associate even if the chipset supports
it but doesn't enable it. So remove the assertion.

Original-Bug: 51985
Tested: Verified on NUC 8265. Associated with an open network.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562045
Original-Revision: cc1349f6b7eec8316562dba0fed4f64fc079a668
GitOrigin-RevId: e40e1934d428cacd91de6cbbddabdf7ba3c2976a
Change-Id: Iadff0e46b487a22a8a8ecce569c04c62e91f8648