[roll] Roll fuchsia [blobfs] Remove BlobLoader::LoadResult.
Remove the BlobLoader::LoadResult and put all information in the
LoaderInfo struct.

The main structure change is that a BlobLayout object is now kept for
every blob. Before the BlobLayout was used to compute a few values which
were saved, and then it was discarded. The BlobLayout struct is
relatively small, this allows less duplication, and it seems like the
values it has should be readily available anyway. This allows some
cleanup in various places.

Original-Bug: 79611
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562143
Original-Revision: 8a764a0b21dcd0629748c186ba6dd0f0175c0084
GitOrigin-RevId: 07f026ff3b8421295172517a91337d1c906920a3
Change-Id: I90a132591c0829c668003fc7b4301387eccf37ef