[roll] Roll experiences [ermine] Handle invalid focus handle from focus chain
Due to a recent change: https://fxrev.dev/561821 (now reverted) the
focus service encountered invalid focus chain. The shell should always
be in the focus chain. But it was receiving a focus chain with only the
root view in it. This results in an invalid app state and an
exception because settings start was called consecutively.

This change adds a guard to only respond when the focus is changing,
basically handle invalid focus chains.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/562644
Original-Revision: d0333bf0722ed22690eef52e3b38fa57f7b60e46
GitOrigin-RevId: 18b05140d86931b203298df045b2b6fad95c8c43
Change-Id: I98ee151ecb3c14d851d683839af4e4f960782219