[roll] Roll fuchsia [ffx][test] Remove experimental flag from ffx test
ffx test is intended to replace ffx component test. While ffx test is
not completely ready, the parts of it that replace the functionality of
ffx component test are ready. This does a few things with the goal of
encouraging users to migrate to ffx test while marking the areas which
are still experimental:
* Add a deprecation notice to ffx component test pointing at ffx test
* Remove the experimental flag on ffx test
* Add an experimental flag around features of ffx test that rely on
the unstable structured test output

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550828
Original-Revision: 786c4c26cef0f71fbc17e6a9f7f3cc55b57fdcc3
GitOrigin-RevId: 39236f066f41f2e89db59730d59f5f86c9ea1dbe
Change-Id: Id2931e41a701620c2161935ad183524354d25cb1