[roll] Roll fuchsia [blobfs] Make reserving blocks more efficient
Skip past all allocated blocks in a region in FindUnallocatedExtent
when trying to find free blocks. Previously if we were trying to
reserve 1 block and the first free block was the 14000th (120MB used
with no holes) then it would take 14000 calls to FindUnallocatedExtent
and 28000 calls to RawBitmap::Scan to find it. This was because
FindUnallocatedExtent would only advance the number of blocks we were
looking for on each call. Now FindUnallocatedExtent advances entire
allocated regions in a single call.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562161
Original-Revision: e265503916db8b73e3c7649bd8e314e07ef5acda
GitOrigin-RevId: 3c8fbacac35d7d112e0f090179cb3c486eb07173
Change-Id: I70385f1e698457c5d59f904d5d5a2c5e81c75e17