[roll] Roll fuchsia [component_manager] don't fail to destroy an instance with bad storage
Failing to route a storage capability would cause instance destruction
to fail in component manager, because it was trying to delete the
instance's storage. If an instance has an invalid route for a storage
capability, then it was never able to write to storage in the first
place, and component manager can safely ignore the error.

Original-Fixed: 81289
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559201
Original-Revision: e6b7551dc19edfc811b1b67ca67422f462e30091
GitOrigin-RevId: 8cae89a22cf0394bccf91d3f3aca3ae8214e0a33
Change-Id: Id4808a8a8f376c706a8ca187e43cd1cc226eb6d3