[roll] Roll fuchsia [debugger] Enhance cross-platform errors.
Adds some non-system-derived error values to the cross-platform
debug::Status struct for the types of errors we expect the frontend
needs to handle specially. This also gives a direct mapping to the
frontend's Err value.

This brings debug::Status and zxdb::Err closer, we may want to consider
merging these in the future.

Currently these values are just derived from the zx_status_t values at
the time of creation so there should be no effective change. But by
removing checks for specific zx_status_t values in the frontend we have
more flexibility in being precise in the backend implementation and not
relying on potentially changing parts of the system.

As part of testing, I noticed that all "run" commands hang on Fuchsia so
some better error messaging is added in this case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562812
Original-Revision: 53e427acb7b0982784dfddf9776d1c563ee26304
GitOrigin-RevId: fb8c20a2c4af4f7aeaf2c074910bfda517aba7fe
Change-Id: I468203f3e659d5852198f606a2f6acb8d450184d