[roll] Roll infra/recipes [fuchsia] Always set *gcs_bucket output props
...even when the spec is passed via properties, rather than loaded from
a file in integration.git.

This is a bug that I introduced in fxrev.dev/539061. That change was
intended to only skip the spec-loading logic if the spec was included in
the input properties, but I accidentally skipped the logic to set
`artifact_gcs_bucket` and `gcs_bucket` output properties as well. That
causes OTA tests that rely on those properties to fail when run on
builders with specs passed via properties.

This also restores the "load spec" step even when the spec comes from
properties, since it's a convenient `presentation` object to use for
emitting the output properties. However, it's also useful to have that
step even if the spec comes from properties, to make it easier for users
to view the spec.

Original-Bug: 77879
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/563392
Original-Revision: b46c1d455a65262ccbedf5a560793a2bf1d19780
GitOrigin-RevId: b77300b6c3e967da2e6e4068374695f85b8c5971
Change-Id: Ie18e58eecbbf302c6fb8d97fffb4a24e2bdba866