[roll] Roll fuchsia [build][sysmgr] Allow duplicate sysmgr config files.
It is illegal to specify a build configuration where multiple
sysmgr config files define the provider of a single service.
However, it's not hard to inadvertently create a situation
where the *identical* config file appears twice.  Instead of
dying with a confusing build error, this CL detects and allows
this situation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554969
Original-Revision: e2732351ebc558660e451e26b441b0db2cafd695
GitOrigin-RevId: b6af96508cd094d070d763026d203b1de2146e0c
Change-Id: Ic1b7a3bdbabb5786c2c3500768e2d006f4359e84