[roll] Roll fuchsia [rust][ffi] Add build templates to use 'cxx' crate
This adds a 'rust_cxx_ffi_library' build template which produces a
source_set linked to a Rust static library with FFI generated using the
'cxx' crate. The template also includes a normal rustc_library target so
that other Rust targets can depend on it as they would elsewhere,
allowing multiple libraries that use 'cxx' to be composed.

cargo-gnaw is also updated to use public_deps instead of deps for the
top-level group targets of binaries. This is required to use the
executable through the group in a compiled_action() target.

Original-Bug: 34560
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420319
Original-Revision: e71d7513668592cd643d6b90258cab1f8cbe2057
GitOrigin-RevId: 6111daf5586033d3e97fbc0aae5523f2101b62d9
Change-Id: I22817ad70d599dc339b5595c08da22cd7fc56a88