[roll] Roll fuchsia [starnix] Pass mmap/truncate tests
After this CL, we pass three tests that combine mmap and ftruncate.
Previously, we were not zeroing out the memory at the end of a file
correctly. Turns out zx_vmo_set_size will do this work for us.

Also, implement a trivial msync that is enough to pass the test. When
Zircon has a writeback pager, we'll be able to have a more complete
implementation that actually triggers a writeback for paged memory.

Test: MMapFileTest.*Truncate*
Original-Bug: 79301
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562908
Original-Revision: 770618eb406ba08f5367953e9ef36b7ddfcecdc6
GitOrigin-RevId: 15363d1570bec89909569fe714fd75194ebcbe7f
Change-Id: I9027b4a15f3bd7ff42380b6a80ae4eb8b57ed5f5